Module shadow_rs::core

The core infrastructure needed to configure and run the simulator.

Modules

configuration

Shadow’s configuration and cli parsing code using serde and clap. This contains all of Shadow’s configuration options, some of which are also exposed as CLI options.

controller

cpu

logger

manager

resource_usage

Utilities for getting system resource usage.

runahead

sim_config

Code for processing parsed Shadow configurations. This involves loading and verifying network graphs, converting options to types/formats that are easier to use in Shadow, verifying that paths exist, etc.

sim_stats

work

worker