Struct shadow_rs::core::configuration::GeneralOptions

pub struct GeneralOptions {
    pub stop_time: Option<Time<TimePrefix>>,
    pub seed: Option<u32>,
    pub parallelism: Option<u32>,
    pub bootstrap_end_time: Option<Time<TimePrefix>>,
    pub log_level: Option<LogLevel>,
    pub heartbeat_interval: Option<NullableOption<Time<TimePrefix>>>,
    pub data_directory: Option<String>,
    pub template_directory: Option<NullableOption<String>>,
    pub progress: Option<bool>,
    pub model_unblocked_syscall_latency: Option<bool>,
}

Fields
stop_time: Option<Time<TimePrefix>>

The simulated time at which simulated processes are sent a SIGKILL signal

seed: Option<u32>

Initialize randomness using seed N

parallelism: Option<u32>

How many parallel threads to use to run the simulation. A value of 0 will allow Shadow to choose the number of threads.

bootstrap_end_time: Option<Time<TimePrefix>>

The simulated time that ends Shadow’s high network bandwidth/reliability bootstrap period

log_level: Option<LogLevel>

Log level of output written on stdout. If Shadow was built in release mode, then log messages at level ‘trace’ will always be dropped

heartbeat_interval: Option<NullableOption<Time<TimePrefix>>>

Interval at which to print heartbeat messages

data_directory: Option<String>

Path to store simulation output

template_directory: Option<NullableOption<String>>

Path to recursively copy during startup and use as the data-directory

progress: Option<bool>

Show the simulation progress on stderr

model_unblocked_syscall_latency: Option<bool>

Model syscalls and VDSO functions that don’t block as having some latency. This should have minimal effect on typical simulations, but can be helpful for programs with “busy loops” that otherwise deadlock under Shadow.

Implementations

impl GeneralOptions

pub fn with_defaults(self, default: Self) -> Self

Replace unset (None) values of base with values from default.