Module shadow_rs::core::cpu

Structs

RangeListIter

Functions

core

Get the core ID from /sys/devices/system/cpu/cpu{cpu}/topology/core_id.

count_physical_cores

Count the number of physical cores available. Uses sched_getaffinity so should take into account CPU affinity and cgroups.

cpus

Get the CPUs in a node from /sys/devices/system/node/node{node}/cpulist.

nodes

Get the nodes from /sys/devices/system/node/possible.

online

Get the online CPUs from /sys/devices/system/cpu/online.

parse_range_list

Take an input of a list of ranges like ‘1-3,5,7-10’ and return an iterator of integers like [1,2,3,5,7,8,9,10]. The returned iterator will panic if the input is not nicely formatted (no whitespace, etc) or contains invalid characters.