Module shadow_rs::core::logger::log_wrapper

Functions

c_to_rust_log_level

optional_str⚠
Returns the str pointed to by ptr if it’s non-NULL and points to a UTF-8 null-terminated string. SAFETY: ptr must point a NULL-terminated C String if non-null, and must be immutable for the lifetime of the returned str.

rustlogger_flush⚠
Flush Rust’s log::logger().

rustlogger_isEnabled⚠
Whether logging is currently enabled for level.

rustlogger_log⚠
Log to Rust’s log::logger().

rustlogger_setLevel⚠
Set the max (noisiest) logging level to level.