Module shadow_rs::core::support::configuration

Structs

CliOptions

ConfigFileOptions

Options contained in a configuration file.

ConfigOptions

Shadow configuration options after processing command-line and configuration file options.

ExperimentalOptions

FileSource

GeneralOptions

HostDefaultOptions

HostOptions

NetworkOptions

ProcessOptions

Quantity

Enums

Compression

GraphOptions

GraphSource

InterposeMethod

LogInfoFlag

LogLevel

NullableOption

This wrapper type allows cli options to specify “null” to overwrite a config file option with None, and is intended to be used for options where “null” is a valid option value.

ProcessArgs

QDiscMode

SchedulerPolicy

StraceLoggingMode

Constants

ONE_GBIT_SWITCH_GRAPH

Functions

tilde_expansion