Module shadow_rs::core::support::configuration::export

Functions

clioptions_freeString

clioptions_getConfig

clioptions_getGdb

clioptions_getShmCleanup

clioptions_getShowBuildInfo

clioptions_getShowConfig

config_freeString

config_getBootstrapEndTime

config_getDataDirectory

config_getHeartbeatInterval

config_getHostHeartbeatInterval

config_getHostHeartbeatLogInfo

config_getHostHeartbeatLogLevel

config_getInterfaceBuffer

config_getInterfaceQdisc

config_getInterposeMethod

config_getLogLevel

config_getNHosts

config_getParallelism

config_getPreloadSpinMax

config_getProgress

config_getRunahead

config_getSchedulerPolicy

config_getSeed

config_getSocketRecvAutotune

config_getSocketRecvBuffer

config_getSocketSendAutotune

config_getSocketSendBuffer

config_getStopTime

config_getTemplateDirectory

config_getUseCpuPinning

config_getUseDynamicRunahead

config_getUseExplicitBlockMessage

config_getUseLegacyWorkingDir

config_getUseLibcPreload

config_getUseMemoryManager

config_getUseObjectCounters

config_getUseOnWaitpidWorkarounds

config_getUseOpensslCryptoPreload

config_getUseOpensslRNGPreload

config_getUseSchedFifo

config_getUseSeccomp

config_getUseShimSyscallHandler

config_getUseShortestPath

config_getUseStraceLogging

config_getUseSyscallCounters

config_getWorkers

config_iterHosts

config_showConfig

hostoptions_freeString

hostoptions_getBandwidthDown

Get the downstream bandwidth of the host if it exists. A non-zero return value means that the host did not have a downstream bandwidth and that bandwidth_down was not updated.

hostoptions_getBandwidthUp

Get the upstream bandwidth of the host if it exists. A non-zero return value means that the host did not have an upstream bandwidth and that bandwidth_up was not updated.

hostoptions_getIpAddr

hostoptions_getLogLevel

hostoptions_getNetworkNodeId

hostoptions_getPcapDirectory

hostoptions_getQuantity

hostoptions_iterProcesses

processoptions_freeString

processoptions_getArgs

processoptions_getEnvironment

processoptions_getPath

Will return a NULL pointer if the path does not exist.

processoptions_getQuantity

processoptions_getRawPath

Returns the path exactly as specified in the config. Caller must free returned string.

processoptions_getStartTime

processoptions_getStopTime