Struct shadow_rs::core::support::configuration::GeneralOptions

pub struct GeneralOptions {
    stop_time: Option<Time<TimePrefixUpper>>,
    seed: Option<u32>,
    parallelism: Option<NonZeroU32>,
    bootstrap_end_time: Option<Time<TimePrefixUpper>>,
    log_level: Option<LogLevel>,
    heartbeat_interval: Option<Time<TimePrefixUpper>>,
    data_directory: Option<String>,
    template_directory: Option<String>,
}

Fields
stop_time: Option<Time<TimePrefixUpper>>

The simulated time at which simulated processes are sent a SIGKILL signal

seed: Option<u32>

Initialize randomness using seed N

parallelism: Option<NonZeroU32>

How many parallel threads to use to run the simulation. Optimal performance is usually obtained with nproc, or sometimes nproc/2 with hyperthreading.

bootstrap_end_time: Option<Time<TimePrefixUpper>>log_level: Option<LogLevel>

Log level of output written on stdout. If Shadow was built in release mode, then log messages at level ‘trace’ will always be dropped

heartbeat_interval: Option<Time<TimePrefixUpper>>

Interval at which to print heartbeat messages

data_directory: Option<String>

Path to store simulation output

template_directory: Option<String>

Path to recursively copy during startup and use as the data-directory

Implementations

impl GeneralOptions

pub fn with_defaults(self, default: Self) -> Self

Replace unset (None) values of base with values from default.