Module shadow_rs::core::support::simulation_time

cbindgen:ignore Values for working with time in simulation units.

In Rust, use EmulatedTime to represent an instant in time, or std::time::Duration to represent a time interval. Use SimulationTime only when interacting with C APIs that use [c::SimulationTime].

This module contains some identically-named constants defined as C macros in main/core/support/definitions.h.

Structs

SimulationTime

Constants

SIMTIME_INVALID

Invalid simulation time.

SIMTIME_MAX

Maximum and minimum valid values.

SIMTIME_MIN

SIMTIME_ONE_HOUR

Represents one hour in simulation time.

SIMTIME_ONE_MICROSECOND

Represents one microsecond in simulation time.

SIMTIME_ONE_MILLISECOND

Represents one millisecond in simulation time.

SIMTIME_ONE_MINUTE

Represents one minute in simulation time.

SIMTIME_ONE_NANOSECOND

Represents one nanosecond in simulation time.

SIMTIME_ONE_SECOND

Represents one second in simulation time.