Module shadow_rs::core::work::task::export

Functions

taskref_drop⚠

Destroys this reference to the Task, dropping the Task if no references remain.

taskref_new_bound⚠

Create a new reference-counted task that can only be executed on the given host. The callbacks can safely assume that they will only be called with the lock for the specified host held.

taskref_new_unbound⚠

Create a new reference-counted task that may be executed on any Host.

Type Aliases

TaskArgumentFreeFunc

TaskCallbackFunc

TaskObjectFreeFunc