shadow_rs::fmt

Struct NumBuffer

pub struct NumBuffer<T>
where
    T: NumBufferTrait,
{ /* private fields */ }

🔬This is a nightly-only experimental API. (int_format_into)

A buffer wrapper of which the internal size is based on the maximum number of digits the associated integer can have.

Examples

#![feature(int_format_into)]
use core::fmt::NumBuffer;

let mut buf = NumBuffer::new();
let n1 = 1972u32;
assert_eq!(n1.format_into(&mut buf), "1972");

// Formatting a negative integer includes the sign.
let mut buf = NumBuffer::new();
let n2 = -1972i32;
assert_eq!(n2.format_into(&mut buf), "-1972");

Implementations

impl<T> NumBuffer<T>
where T: NumBufferTrait,

pub const fn new() -> NumBuffer<T>

🔬This is a nightly-only experimental API. (int_format_into)

Initializes internal buffer.

pub const fn capacity(&self) -> usize

🔬This is a nightly-only experimental API. (int_format_into)

Returns the length of the internal buffer.