Module shadow_rs::host

Modules

context

This module provides several Context structs, intended to bundle together current relevant objects in the hierarchy. These are meant to replace worker_getActiveThread, etc. Passing around the current context explicitly instead of putting them in globals both allows us to avoid interior mutability (and its associated runtime cost and fallibility), and lets us keep a hierarchical object structure (e.g. allow holding a mutable Process and a mutable Thread belonging to that process simultaneously).

descriptor

host

memory_manager

Access and manage memory of a plugin process.

process

syscall

syscall_condition

syscall_types

thread

timer