Module shadow_rs::host::descriptor

Modules

descriptor_table

eventfd

pipe

shared_buf

socket

Structs

Descriptor

DescriptorFlags

FileMode

These are flags that should generally not change (analagous to the Linux filp->f_mode). Since the plugin will never see these values and they’re not exposed by the kernel, we don’t match the kernel FMODE_ values here.

FileState

FileStatus

These are flags that can potentially be changed from the plugin (analagous to the Linux filp->f_flags status flags). Not all O_ flags are valid here. For example file access mode flags (ex: O_RDWR) are stored elsewhere, and file creation flags (ex: O_CREAT) are not stored anywhere. Many of these can be represented in different ways, for example: O_NONBLOCK, SOCK_NONBLOCK, EFD_NONBLOCK, GRND_NONBLOCK, etc, and not all have the same value.

OwnedLegacyDescriptor

Represents an owned reference to a legacy descriptor. Will decrement the descriptor’s ref count when dropped.

Enums

CompatDescriptor

PosixFile

Represents a POSIX description, or a Linux “struct file”.

PosixFileRef

PosixFileRefMut

StateListenerFilter