Enum shadow_rs::host::descriptor::socket::inet::InetSocket

pub enum InetSocket {
    LegacyTcp(Arc<AtomicRefCell<LegacyTcpSocket>>),
    Tcp(Arc<AtomicRefCell<TcpSocket>>),
    Udp(Arc<AtomicRefCell<UdpSocket>>),
}

Variants

LegacyTcp(Arc<AtomicRefCell<LegacyTcpSocket>>)

Tcp(Arc<AtomicRefCell<TcpSocket>>)

Udp(Arc<AtomicRefCell<UdpSocket>>)

Implementations

impl InetSocket

pub fn borrow(&self) -> InetSocketRef<'_>

pub fn try_borrow(&self) -> Result<InetSocketRef<'_>, BorrowError>

pub fn borrow_mut(&self) -> InetSocketRefMut<'_>

pub fn try_borrow_mut(&self) -> Result<InetSocketRefMut<'_>, BorrowMutError>

pub fn downgrade(&self) -> InetSocketWeak

pub fn canonical_handle(&self) -> usize

Useful for getting a unique integer handle for a socket, or when we need to compare a C LegacySocket to a rust InetSocket (which may internally point to the same LegacySocket).

pub fn bind( &self, addr: Option<&SockaddrStorage>, net_ns: &NetworkNamespace, rng: impl Rng, ) -> SyscallResult

pub fn listen( &self, backlog: i32, net_ns: &NetworkNamespace, rng: impl Rng, cb_queue: &mut CallbackQueue, ) -> Result<(), SyscallError>

pub fn connect( &self, addr: &SockaddrStorage, net_ns: &NetworkNamespace, rng: impl Rng, cb_queue: &mut CallbackQueue, ) -> Result<(), SyscallError>

pub fn sendmsg( &self, args: SendmsgArgs<'_>, memory_manager: &mut MemoryManager, net_ns: &NetworkNamespace, rng: impl Rng, cb_queue: &mut CallbackQueue, ) -> Result<ssize_t, SyscallError>

pub fn recvmsg( &self, args: RecvmsgArgs<'_>, memory_manager: &mut MemoryManager, cb_queue: &mut CallbackQueue, ) -> Result<RecvmsgReturn, SyscallError>