shadow_rs::host

Module fcntl_lock_table

Structs

FcntlLockTable

All record locks for a Host.

FileId

Stable identifier for a file, which we use to look up record locks. Remains valid as long as the file has links in the file system and/or there are open file descriptors to the file.

Enums

LockError

Error returned by a lock operation.

LockOwner

Owner of a record lock.