Struct shadow_rs::host::host::Host

pub struct Host { /* private fields */ }

A simulated Host.

This is currently an ephemeral proxy object a C Host (cshadow::Host). Eventually cshadow::Host’s contents and functionality will be migrated into there though, and this will become the “real” Host object.

Implementations

impl Host

pub unsafe fn borrow_from_c(p: *mut Host) -> Self

For now, this should only be called via Worker, to borrow the current Host. This will ensure there is only one reference to a given Host in Rust.

SAFETY: p must point to a valid c::Host, to which this Host will have exclusive access over its lifetime. p must outlive the returned object.

pub fn info(&self) -> &Arc<HostInfo>

Information about the Host. Made available as an Arc for cheap cloning into, e.g. Worker and ShadowLogger. When there’s no need to clone the Arc, generally prefer the top-level Host methods for accessing this information, which are likely to be more stable.

pub fn id(&self) -> HostId

pub fn name(&self) -> &str

pub fn default_ip(&self) -> IpAddr

pub fn abstract_unix_namespace(
&self
) -> &Arc<AtomicRefCell<AbstractUnixNamespace>>

pub fn log_level(&self) -> Option<LevelFilter>

pub fn random(&mut self) -> &mut impl Rng

pub fn get_new_event_id(&mut self) -> u64

pub fn continue_execution_timer(&mut self)

pub fn stop_execution_timer(&mut self)

pub fn schedule_task_at_emulated_time(
&mut self,
task: TaskRef,
t: EmulatedTime
) -> bool

pub fn schedule_task_with_delay(
&mut self,
task: TaskRef,
t: SimulationTime
) -> bool

pub fn event_queue(&self) -> Arc<ThreadSafeEventQueue>

pub fn boot(&mut self)

pub fn shutdown(&mut self)

pub fn free_all_applications(&mut self)

pub fn execute(&mut self, until: EmulatedTime)

pub fn next_event_time(&self) -> Option<EmulatedTime>

pub unsafe fn lock(&mut self)

pub unsafe fn unlock(&mut self)

pub fn chost(&self) -> *mut Host