Module shadow_rs::host::memory_manager::export

Functions

allocdmem_free⚠

allocdmem_new⚠

allocdmem_pluginPtr⚠

memorymanager_free⚠

Safety

memorymanager_freeMutRefWithFlush⚠

Write-back any previously returned writable memory, and free the writer.

memorymanager_freeMutRefWithoutFlush⚠

Write-back any previously returned writable memory, and free the writer.

memorymanager_freeRef⚠

memorymanager_getMutablePtr⚠

Get a readable and writable pointer to this writer’s memory.

memorymanager_getReadablePtr⚠

Get a read-accessor to the specified plugin memory. Must be freed via memorymanager_freeReader.

memorymanager_getReadablePtrPrefix⚠

memorymanager_getWritablePtr⚠

Get a writable pointer to this writer’s memory. Initial contents are unspecified.

memorymanager_handleBrk⚠

Fully handles the brk syscall, keeping the “heap” mapped in our shared mem file.

memorymanager_handleMmap⚠

Fully handles the mmap syscall

memorymanager_handleMprotect⚠

memorymanager_handleMremap⚠

memorymanager_handleMunmap⚠

Fully handles the munmap syscall

memorymanager_initMapperIfNeeded⚠

Initialize the MemoryMapper if it isn’t already initialized. thread must be running and ready to make native syscalls.

memorymanager_new⚠

Safety

memorymanager_readPtr⚠

Copy data from this reader’s memory.

memorymanager_readString⚠

memorymanager_writePtr⚠

Write data to this writer’s memory.

memorymanagermut_ptr⚠

memorymanagermut_sizeof⚠

memorymanagerref_ptr⚠

memorymanagerref_sizeof⚠