Module shadow_rs::host::memory_manager::memory_mapper

Structs

MemoryMapper
Manages the address-space for a plugin process.

Region

ShmFile
Shared memory file into which we relocate parts of the plugin’s address space.

Constants

HEAP_PROT

STACK_PROT

Functions

get_heap
Find the heap range, and map it if non-empty.

get_regions
Get the current mapped regions of the process.

log_regions

map_stack
Finds where the stack is located and maps the region bounding the maximum stack size.