shadow_rs::host::network
Module interface

Structs

NetworkInterface

Represents a network device that can send and receive packets.

PcapOptions

Type Aliases

FifoPacketPriority

The priority used by the fifo qdisc to choose the next socket to send a packet from.