shadow_rs::host::process

Struct RunnableProcess

pub struct RunnableProcess { /* private fields */ }

A process that is currently runnable.

Implementations

impl RunnableProcess

pub fn spawn_mthread_for_exec( &self, host: &Host, plugin_path: &CStr, argv: Vec<CString>, envv: Vec<CString>, ) -> Result<ManagedThread, Errno>

Spawn a ManagedThread corresponding to the given exec syscall parameters. Intended for use by the exec syscall handlers. Whether it succeeds or fails, does not mutate self, though self’s strace and shim log files will be passed into the new ManagedThread.

In case the native exec syscall fails, the corresponding error is returned.

pub fn free_unsafe_borrows_flush(&self) -> Result<(), Errno>

This cleans up memory references left over from legacy C code; usually a syscall handler.

Writes the leftover mutable ref to memory (if any), and frees all memory refs.

pub fn free_unsafe_borrows_noflush(&self)

This cleans up memory references left over from legacy C code; usually a syscall handler.

Frees all memory refs without writing back to memory.

pub fn memory_borrow(&self) -> impl Deref<Target = MemoryManager> + '_

pub fn memory_borrow_mut(&self) -> impl DerefMut<Target = MemoryManager> + '_

pub fn strace_logging_options(&self) -> Option<FmtOptions>

pub fn with_strace_file<T>(&self, f: impl FnOnce(&mut File) -> T) -> Option<T>

If strace logging is disabled, this function will do nothing and return None.

pub fn native_pid(&self) -> Pid

pub fn first_live_thread_borrow( &self, root: &Root, ) -> Option<impl Deref<Target = RootedRc<RootedRefCell<Thread>>> + '_>

Returns a dynamically borrowed reference to the first live thread. This is meant primarily for the MemoryManager.

pub fn thread_borrow( &self, virtual_tid: ThreadId, ) -> Option<impl Deref<Target = RootedRc<RootedRefCell<Thread>>> + '_>

pub fn signal( &self, host: &Host, current_thread: Option<&Thread>, siginfo_t: &siginfo_t, )

Send the signal described in siginfo to process. current_thread should be set if there is one (e.g. if this is being called from a syscall handler), and None otherwise (e.g. when called from a timer expiration event).

An event will be scheduled to deliver the signal unless current_thread is set, and belongs to the process self, and doesn’t have the signal blocked. In that the signal will be processed synchronously when returning from the current syscall.

pub fn add_thread(&self, host: &Host, thread: RootedRc<RootedRefCell<Thread>>)

Adds a new thread to the process and schedules it to run. Intended for use by clone.

pub fn new_forked_process( &self, host: &Host, flags: CloneFlags, exit_signal: Option<Signal>, new_thread_group_leader: RootedRc<RootedRefCell<Thread>>, ) -> RootedRc<RootedRefCell<Process>>

Create a new Process, forked from self, with the thread new_thread_group_leader.

pub fn shmem( &self, ) -> impl Deref<Target = ShMemBlock<'static, ProcessShmem>> + '_

Shared memory for this process.