Module shadow_rs::host::syscall::format

Structs

NoArg

A marker type for indicating there are no types left in the syscall arguments.

SyscallArgsFmt

A formatting wrapper for six syscall arguments.

SyscallPtr

A typed PluginPtr.

SyscallResultFmt

A formatting wrapper for the syscall result.

Enums

FmtOptions

StraceFmtMode

SyscallVal

A syscall value, either data or a pointer.

Traits

FromSyscallMem

A trait for objects that can be read from plugin memory.

SyscallDataDisplay

Format trait for syscall data.

SyscallPtrDisplay

Format trait for syscall pointers. Should be implemented on SyscallPtr<*const T> or SyscallPtr<[T; K]> types.

TryFromSyscallReg

Convert from a SysCallReg.

Functions

write_syscall

Format and write the syscall.