shadow_rs::host::syscall::formatter

Struct SyscallVal

pub struct SyscallVal<'a, T> {
    pub reg: SyscallReg,
    pub args: [SyscallReg; 6],
    /* private fields */
}

A syscall argument or return value. It implements Display, and only reads memory and converts types when being formatted.

Fields
reg: SyscallRegargs: [SyscallReg; 6]
Implementations

impl<'a, T> SyscallVal<'a, T>

pub fn new( reg: SyscallReg, args: [SyscallReg; 6], options: FmtOptions, mem: &'a MemoryManager, ) -> Self

pub fn cast<V>(&self) -> SyscallVal<'a, V>

Cast a syscall argument or return value to another type.