Module shadow_rs::host::syscall_types

Structs

PluginPtr

TypedPluginPtr

Wrapper around a PluginPtr that encapsulates its type, size, and current position.

Enums

SyscallError

Traits

NoTypeInference

A trait to prevent type inference during function calls. Useful when you have a type that wraps a pointer (like TypedPluginPtr) and you don’t want Rust to infer the type of pointer during creation. Instead, the caller must specify the generic type.

Type Definitions

SysCallArgs

SysCallReg

SyscallResult