shadow_rs::host::timer::export

Function timer_new

#[unsafe(no_mangle)]
pub unsafe extern "C-unwind" fn timer_new(
    task: *const TaskRef,
) -> *mut Timer

Create a new Timer that synchronously executes task on expiration.

Safety

task must be dereferenceable, and must not call mutable methods of the enclosing Timer; if it needs to do so it should schedule a new task to do so.