Struct shadow_rs::host::timer::Timer

pub struct Timer { /* private fields */ }

Implementations

impl Timer

pub fn new<F: 'static + Fn(&Host) + Send + Sync>(on_expire: F) -> Self

Create a new Timer that directly executes on_expire on expiration. on_expire will cause a panic if it calls mutable methods of the enclosing Timer. If it may need to call mutable methods of the Timer, it should push a new task to the scheduler to do so.

pub fn expiration_count(&self) -> u64

Returns the number of timer expirations that have occurred since the last time Timer::consume_expiration_count() was called without resetting the counter.

pub fn expire_interval(&self) -> Option<SimulationTime>

Returns the currently configured timer expiration interval if this timer is configured to periodically expire, or None if the timer is configured for a one-shot expiration.

pub fn consume_expiration_count(&mut self) -> u64

Returns the number of timer expirations that have occurred since the last time Timer::consume_expiration_count() was called and resets the counter to zero.

pub fn remaining_time(&self) -> Option<SimulationTime>

Returns the remaining time until the next expiration if the timer is armed, or None otherwise.

pub fn disarm(&mut self)

Deactivate the timer so that it does not issue on_expire() callback notifications.

pub fn arm( &mut self, host: &Host, expire_time: EmulatedTime, expire_interval: Option<SimulationTime>, )

Activate the timer so that it starts issuing on_expire() callback notifications.

The expire_time instant specifies the next time that the timer will expire and issue an on_expire() notification callback. The expire_interval duration is optional: if Some, it configures the timer in periodic mode where it issues on_expire() notification callbacks every interval of time; if None, the timer is configured in one-shot mode and will become disarmed after the first expiration.

Panics if expire_time is in the past or if expire_interval is Some but not positive.