shadow_rs
Macro log_once_at_level

macro_rules! log_once_at_level {
    ($lvl_once:expr, $lvl_remaining:expr, $str:literal $($x:tt)*) => { ... };
}

Log a message once at level lvl_once, and any later log messages from this line at level lvl_remaining.

A log target is not supported. The string “(LOG_ONCE)” will be prepended to the message to indicate that future messages won’t be logged at lvl_once.

log_once_at_level!(Level::Warn, Level::Debug, "Unexpected flag {}", 10);