shadow_rs::network
Module graph

Structs

IpAssignment

Tool for assigning IP addresses to graph nodes.

IpPreviouslyAssignedError

NetworkGraph

A network graph containing the petgraph graph and a map from gml node ids to petgraph node indexes.

PathProperties

Network characteristics for a path between two nodes.

RoutingInfo

Routing information for paths between nodes.

ShadowEdge

A graph edge.

ShadowNode

A graph node.

Functions

load_network_graph

Get the network graph as a string.