shadow_rs::network
Module packet

Structs

Packet

Holds networking information and payload data that assists in sending managed-process bytes between different sockets and hosts over the simulation network.

PacketRc

A thread-safe shared reference to a Packet.

Enums

IanaProtocol

Official IANA-assigned protocols supported in our packets.

PacketStatus

Represents different checkpoints that a packet reaches as it is being moved around in Shadow.

TypeOfService

A packet’s type of service (TOS) indicates its desired queuing priority. This may be used by Shadow’s interface qdisc to prioritize packets that it sends out to the network. For example, in the pfifo_fast qdisc, packets in lower priority bands would be sent ahead of others.