shadow_rs::network::packet

Struct Packet

pub struct Packet { /* private fields */ }

Holds networking information and payload data that assists in sending managed-process bytes between different sockets and hosts over the simulation network.

The Packet is designed to separate the IP layer header from the transport layer data, to make it easier to support new transport protocols in the future. Thus, the Packet has a Header to store IP header information, and a Data to store transport-specific data (e.g., A TCP header and payload for TCP, or a UDP header and payload for UDP).

The Packet is designed to be read-only after creation to simplify implementation and to reduce the need for mutable borrows.

Deprecation

As an exception to the read-only design, some legacy TCP data is mutable to support the legacy C TCP stack. When the legacy TCP stack and the Packetโs C interface is removed, mutability will also be removed.

Implementations

impl Packet

pub fn new_ipv4_tcp( header: TcpHeader, payload: Payload, priority: FifoPacketPriority, ) -> Self

Creates a new IPv4 TCP packet using the provided data.

pub fn new_ipv4_udp( src: SocketAddrV4, dst: SocketAddrV4, payload: Bytes, priority: FifoPacketPriority, ) -> Self

Creates a new IPv4 UDP packet using the provided data.

pub fn ipv4_tcp_header(&self) -> Option<TcpHeader>

If the packet is an IPv4 TCP packet, returns a copy of the TCP header in a format defined by the Rust TCP stack. Otherwise, returns None.

Panics

This function panics if the packet was created with packet_new_tcp() in the legacy C API.

pub fn payload(&self) -> Vec<Bytes>

Returns the packetโs payload that was provided at packet creation time. This function allocates a new Vec, but is zero-copy with respect to the payload Bytes.

This function may return a non-empty vector of zero-length Bytes object(s) if zero-length Bytes object(s) were provided at creation time. Thus, it may be helpful to check if the packet has useful payload using payload_len() before calling this function.

pub fn len(&self) -> usize

Returns the total simulated length of the packet, which is the sum of the emulated IP and transport header lengths and the payload length.

pub fn payload_len(&self) -> usize

Returns the total number of payload bytes stored in the packet, which excludes IP and transport header lengths.

pub fn add_status(&self, status: PacketStatus)

Appends the provided packet status to the list of the packetโs status checkpoints.

This function has no effect unless log::Level::Trace is enabled.

pub fn src_ipv4_address(&self) -> SocketAddrV4

Returns the packetโs IPv4 source address and source port.

Panics

This function panics if the source address is not an IPv4 address.

pub fn dst_ipv4_address(&self) -> SocketAddrV4

Returns the packetโs IPv4 destination address and destination port.

Panics

This function panics if the destination address is not an IPv4 address.

pub fn priority(&self) -> FifoPacketPriority

Returns the priority set at packet creation time.

pub fn iana_protocol(&self) -> IanaProtocol

Returns the packetโs iana-assigned protocol type.