shadow_rs::network::packet

Struct PacketRc

pub struct PacketRc { /* private fields */ }

A thread-safe shared reference to a Packet.

Although a PacketRc is thread-safe, a panic may occur if multiple threads concurrently write PacketStatus metadata on the inner Packet. Thus, use new_copy_inner() to copy the inner Packet and send the copy to the other thread rather than sharing it between threads.

A PacketRc is a wrapper around a Packet and allows us to support reference counting on Packets while safely sharing them across threads. A clone of a PacketRcs increments the reference count of the shared Packet while dropping the PacketRc decrements the reference count. The shared inner Packet is only dropped when all PacketRcs referring to it are also dropped (i.e., the reference count reaches zero).

The PartialEq implementation on PacketRc compares the pointer values of the wrapped Packet.

PacketRc implements the Deref trait so that all non-associated functions on Packet can be accessed from instances of PacketRc too.

Implementations

impl PacketRc

pub fn new_ipv4_tcp( header: TcpHeader, payload: Payload, priority: FifoPacketPriority, ) -> Self

Creates a thread-safe shared reference to a new Packet using the provided information. Additional references to the Packet can be cheaply obtained by cloning the returned PacketRc. The Packet is dropped when its last PacketRc reference is dropped.

See Packet::new_ipv4_tcp() for more details.

pub fn new_ipv4_udp( src: SocketAddrV4, dst: SocketAddrV4, payload: Bytes, priority: FifoPacketPriority, ) -> Self

Creates a thread-safe shared reference to a new Packet using the provided information. Additional references to the Packet can be cheaply obtained by cloning the returned PacketRc. The Packet is dropped when its last PacketRc reference is dropped.

See Packet::new_ipv4_udp() for more details.

pub fn new_copy_inner(&self) -> Self

Creates a thread-safe shared reference to a new Packet that is created by performing a copy of the provided referenced Packet. This function copies all packet data except the packet payload, which is shared by incrementing its reference count. Additional references to the Packet can be cheaply obtained by cloning the returned PacketRc. The Packet is dropped when its last PacketRc reference is dropped.

pub fn from_raw(packet_ptr: *mut Packet) -> Self

Transfers ownership of the given packet_ptr reference into a new PacketRc object. The provided pointer must have been obtained from a call to the Rust function PacketRc::into_raw() or the C function packet_new_tcp().

Use PacketRc::into_raw() if the returned reference is to be handed back to C code.

Panics

This function panics if the supplied packet pointer is NULL.

Deprecation

This function provides compatibility with the legacy C TCP stack and should be considered deprecated and removed when the legacy C TCP stack is removed.

pub fn into_raw(self) -> *mut Packet

Transfers ownership of the inner Arc reference to the caller while dropping the PacketRc.

To avoid a memory leak, the returned pointer must be either reconstituted into a PacketRc using the Rust function PacketRc::from_raw(), or dropped using the C function packet_unref().

Although this returns a *mut Packet pointer, the packet is not actually mutuable. We return a *mut Packet pointer only to avoid having to change the instnaces of the pointers to const instances in the C network code.

Deprecation

This function provides compatibility with the legacy C TCP stack and should be considered deprecated and removed when the legacy C TCP stack is removed.