Module shadow_rs::utility

Modules

byte_queue
A shared buffer that is composed of several chunks. The buffer can be read and written and guarantees it will not allow reading more than was written. Its basically a linked queue that is written (and grows) at the front and read (and shrinks) from the back. As data is written, new chunks are created automatically. As data is read, old chunks are freed automatically.

counter
A counter that can be used to count frequencies of a set of objects. The counter starts with no keys. When an unknown key is incremented, the counter adds a new key to an internal map and sets the count for that key to 1. When a known key is incremented, the count for that key increases. The state of the counter can be extracted by converting it to a string, which lists the counts for all keys sorted with the heaviest hitters first. Currently, only String types are supported, but we may eventually support counting generic types.

event_queue

interval_map

pod

proc_maps

syscall