Struct shadow_rs::utility::byte_queue::ByteQueue

pub struct ByteQueue {
    chunks: LinkedList<ByteChunk>,
    tail_read_offset: usize,
    length: usize,
    chunk_capacity: usize,
}

A queue of byte chunks.

Fields
chunks: LinkedList<ByteChunk>tail_read_offset: usizelength: usizechunk_capacity: usize
Implementations

impl ByteQueue

pub fn new(chunk_capacity: usize) -> ByteQueue

pub fn len(&self) -> usize

Number of bytes in the queue.

pub fn is_empty(&self) -> bool

pub fn push<R: Read>(&mut self, src: R) -> Result<usize>

Push bytes to the head of the queue. Returns an error iff src returns an error; i.e. this is infallible for an infallible src such as a slice.

pub fn pop<W: Write>(&mut self, dst: W) -> Result<usize>

Dequeues data into dst until empty or dst returns EOF or error. Returns an error iff dst returns an error; i.e. this is infallible for an infallible dst such as a slice.

fn create_new_head(&mut self)

fn destroy_old_tail(&mut self)

fn get_available_bytes_tail(&self) -> usize