Struct shadow_rs::utility::callback_queue::CallbackQueue

pub struct CallbackQueue(/* private fields */);

A queue of events (functions/closures) which when run can add their own events to the queue. This allows events to be deferred and run later.

Implementations

impl CallbackQueue

pub fn new() -> Self

Create an empty event queue.

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn add(&mut self, f: impl FnOnce(&mut Self) + 'static)

Add an event to the queue.

pub fn run(&mut self)

Process all of the events in the queue (and any new events that are generated).

pub fn queue_and_run_with_legacy<F, U>(f: F) -> U
where F: FnOnce(&mut Self) -> U,

A convenience function to create a CallbackQueue, allow the caller to add events, and process them all before returning.

This also has the side-effect of ensuring that a global thread-local queue is configured for C code using with_global_cb_queue. We do this for convenience and to help prevent bugs where we forget to call with_global_cb_queue. Ideally we’d like to remove this side-effect as we remove more C code from Shadow.

TODO: Once we have removed C file objects, remove this function and make Self::queue_and_run public.