Function shadow_rs::utility::childpid_watcher::export::childpidwatcher_registerPid

#[no_mangle]
pub unsafe extern "C" fn childpidwatcher_registerPid(
    watcher: *const ChildPidWatcher, 
    pid: i32, 
    read_fd: i32
)

Register interest in pid, and associate it with read_fd.

read_fd should be the read end of a pipe, whose write end is owned solely by pid, causing read_fd to become invalid when pid exits. In a multi-threaded program care must be taken to prevent a concurrent fork from leaking the write end of the pipe into other children. One way to avoid this is to use O_CLOEXEC when creating the pipe, and then unset O_CLOEXEC in the child before calling exec.

Be sure to close the parent’s write-end of the pipe.

Takes ownership of read_fd, and will close it when appropriate.