Function shadow_rs::utility::childpid_watcher::export::childpidwatcher_watch

#[no_mangle]
pub unsafe extern "C" fn childpidwatcher_watch(
    watcher: *const ChildPidWatcher, 
    pid: pid_t, 
    callback: extern "C" fn(_: pid_t, _: *mut c_void), 
    data: *mut c_void
) -> WatchHandle

Call callback exactly once from another thread after the child pid has exited, including if it has already exited. Does not reap the child itself.

The returned handle is guaranteed to be non-zero.

Panics if pid doesn’t exist.