Struct shadow_rs::utility::childpid_watcher::ChildPidWatcher

pub struct ChildPidWatcher {
    inner: Arc<Mutex<Inner>>,
    epoll: RawFd,
}

Utility for monitoring a set of child pid’s, calling registered callbacks when one exits or is killed. Starts a background thread, which is shut down when the object is dropped.

Fields
inner: Arc<Mutex<Inner>>epoll: RawFd
Implementations

impl ChildPidWatcher

pub fn new() -> Self

Create a ChildPidWatcher. Spawns a background thread, which is joined when the object is dropped.

fn thread_loop(inner: &Mutex<Inner>, epoll: RawFd)

unsafe fn fork_watchable_internal(
&self,
fork_syscall: i64,
child_fn: impl FnOnce()
) -> Result<Pid, Error>

pub unsafe fn fork_watchable(
&self,
child_fn: impl FnOnce()
) -> Result<Pid, Error>

Fork a child and register it. Uses fork internally; it vfork is desired, use register_pid instead.

TODO: add a vfork version when Rust supports vfork: https://github.com/rust-lang/rust/issues/58314

pub fn register_pid(&self, pid: Pid, read_fd: File)

Register interest in pid, and associate it with read_fd.

read_fd should be the read end of a pipe, whose write end is owned solely by pid, causing read_fd to become invalid when pid exits. In a multi-threaded program care must be taken to prevent a concurrent fork from leaking the write end of the pipe into other children. One way to avoid this is to use O_CLOEXEC when creating the pipe, and then unset O_CLOEXEC in the child before calling exec.

Be sure to close the parent’s write-end of the pipe.

Takes ownership of read_fd, and will close it when appropriate.

pub fn unregister_pid(&self, pid: Pid)

Unregister the pid. After unregistration, no more callbacks may be registered for the given pid. Already-registered callbacks will still be called if and when the pid exits unless individually unregistered.

Safe to call multiple times.

pub fn register_callback(
&self,
pid: Pid,
callback: impl Send + FnOnce(Pid) + 'static
) -> WatchHandle

Call callback from another thread after the child pid has exited, including if it has already exited. Does not reap the child itself.

The returned handle is guaranteed to be non-zero.

Panics if pid isn’t registered.

pub fn unregister_callback(&self, pid: Pid, handle: WatchHandle)

Unregisters a callback. After returning, the corresponding callback is guaranteed either to have already run, or to never run. i.e. it’s safe to free data that the callback might otherwise access.

Panics if pid was never registered.