shadow_rs::utility::interval_map

Struct IntervalMap

pub struct IntervalMap<V> { /* private fields */ }

Implementations

impl<V: Clone> IntervalMap<V>

Maps from non-overlapping Intervals to V.

pub fn new() -> IntervalMap<V>

pub fn keys(&self) -> KeyIter<'_, V>

Returns iterator over all intervals keys, in sorted order.

pub fn iter(&self) -> ItemIter<'_, V>

Returns iterator over all intervals keys and their values, in order by interval key.

pub fn iter_from(&self, begin: usize) -> ItemIter<'_, V>

Returns iterator over all interval keys and their values, starting with the first interval containing or after begin.

pub fn clear(&mut self, interval: Interval) -> Vec<Mutation<V>>

Mutates the map so that the given range maps to nothing, modifying and removing intervals as needed. Returns what mutations were performed, including the values of any completely-removed intervals. If an interval is split (e.g. by inserting [5,6] into [0,10]), its value will be cloned.

Returned mutations are ordered by their original interval values.

pub fn insert(&mut self, interval: Interval, val: V) -> Vec<Mutation<V>>

Insert range from start to end, inclusive, mapping that range to val. Existing contents of that range are cleared, and mutations returned, as for clear.

pub fn get(&self, x: usize) -> Option<(Interval, &V)>

pub fn get_mut(&mut self, x: usize) -> Option<(Interval, &mut V)>