Struct shadow_rs::utility::sockaddr::SockaddrUnix

pub struct SockaddrUnix<T>
where
    T: Borrow<sockaddr_un>,
{ /* private fields */ }

A Unix socket address. Typically will be used as an owned address SockaddrUnix<libc::sockaddr_un> or a borrowed address SockaddrUnix<&libc::sockaddr_un>, and you can convert between them using methods such as as_ref or into_owned.

Implementations

impl<T> SockaddrUnix<T>
where T: Borrow<sockaddr_un>,

pub fn new(addr: T, len: socklen_t) -> Option<Self>

Get a new SockaddrUnix for a libc::sockaddr_un. The libc::sockaddr_un must be properly initialized. Will return None if the length is too short, too long, or if sun_family is not AF_UNIX.

pub fn as_path(&self) -> Option<&CStr>

If the socket address represents a pathname address, returns the C string representing the filesystem path.

pub fn as_abstract(&self) -> Option<&[u8]>

If the socket address represents an abstract address, returns the bytes representing the name of the abstract socket address. These bytes do not include the nul byte at sun_path[0].

pub fn is_unnamed(&self) -> bool

Is the unix socket address unnamed? On Linux, unnamed unix sockets are unnamed if their length is size_of::<libc::sa_family_t>().

pub fn into_owned(self) -> SockaddrUnix<sockaddr_un>

Get an owned unix socket address.

pub fn as_ref(&self) -> SockaddrUnix<&sockaddr_un>

Get a borrowed unix socket address.

pub fn as_ptr(&self) -> (*const sockaddr_un, socklen_t)

Get a pointer to the unix socket address. All fields of the libc::sockaddr_un will be properly initialized.

impl<T> SockaddrUnix<T>
where T: BorrowMut<sockaddr_un>,

pub fn as_mut(&mut self) -> SockaddrUnix<&mut sockaddr_un>

Get a mutably borrowed unix socket address.

impl SockaddrUnix<sockaddr_un>

pub fn new_path(path: &CStr) -> Option<Self>

Get a new SockaddrUnix with the given path. Will return None if the path is empty or is too large.

pub fn new_abstract(name: &[u8]) -> Option<Self>

Get a new SockaddrUnix with the given abstract address name. The name does not include the required nul byte at sun_path[0]. Will return None if the name is too large.

pub fn new_unnamed() -> SockaddrUnix<sockaddr_un>

Get a new unnamed unix socket address.