Crate shadow_rs

The Shadow network simulator.

Shadow is a discrete-event network simulator that directly executes real application code, enabling you to simulate distributed systems with thousands of network-connected processes in realistic and scalable private network experiments using your laptop, desktop, or server running Linux.

Modules

core

The core infrastructure needed to configure and run the simulator.

cshadow

cbindgen:ignore

host

Process management and emulation of Linux behaviour.

network

The network simulation.

shadow

The main entrypoint to Shadow.

utility

Miscellaneous utilities that are used by Shadow.

Macros

log_once_at_level

Log a message once at level lvl_once, and any later log messages from this line at level lvl_remaining.

log_once_per_value_at_level

Log a message once at level lvl_once for each distinct value, and any later log messages from this line with an already-logged value at level lvl_remaining.