Struct shadow_rs::core::configuration::ExperimentalOptions

pub struct ExperimentalOptions {

    pub use_sched_fifo: Option<bool>,
    pub use_syscall_counters: Option<bool>,
    pub use_object_counters: Option<bool>,
    pub use_preload_libc: Option<bool>,
    pub use_preload_openssl_rng: Option<bool>,
    pub use_preload_openssl_crypto: Option<bool>,
    pub use_memory_manager: Option<bool>,
    pub use_cpu_pinning: Option<bool>,
    pub use_worker_spinning: Option<bool>,
    pub runahead: Option<NullableOption<Time<TimePrefix>>>,
    pub use_dynamic_runahead: Option<bool>,
    pub socket_send_buffer: Option<Bytes<SiPrefixUpper>>,
    pub socket_send_autotune: Option<bool>,
    pub socket_recv_buffer: Option<Bytes<SiPrefixUpper>>,
    pub socket_recv_autotune: Option<bool>,
    pub interface_qdisc: Option<QDiscMode>,
    pub host_heartbeat_log_level: Option<LogLevel>,
    pub host_heartbeat_log_info: Option<HashSet<LogInfoFlag>>,
    pub host_heartbeat_interval: Option<NullableOption<Time<TimePrefix>>>,
    pub strace_logging_mode: Option<StraceLoggingMode>,
    pub max_unapplied_cpu_latency: Option<Time<TimePrefix>>,
    pub unblocked_syscall_latency: Option<Time<TimePrefix>>,
    pub unblocked_vdso_latency: Option<Time<TimePrefix>>,
    pub scheduler: Option<Scheduler>,
    pub log_errors_to_tty: Option<bool>,
    pub use_new_tcp: Option<bool>,

}

Fields
use_sched_fifo: Option<bool>

Use the SCHED_FIFO scheduler. Requires CAP_SYS_NICE. See sched(7), capabilities(7)

use_syscall_counters: Option<bool>

Count the number of occurrences for individual syscalls

use_object_counters: Option<bool>

Count object allocations and deallocations. If disabled, we will not be able to detect object memory leaks

use_preload_libc: Option<bool>

Preload our libc library for all managed processes for fast syscall interposition when possible.

use_preload_openssl_rng: Option<bool>

Preload our OpenSSL RNG library for all managed processes to mitigate non-deterministic use of OpenSSL.

use_preload_openssl_crypto: Option<bool>

Preload our OpenSSL crypto library for all managed processes to skip some crypto operations (may speed up simulation if your CPU lacks AES-NI support, but can cause bugs so do not use unless you know what you’re doing).

use_memory_manager: Option<bool>

Use the MemoryManager in memory-mapping mode. This can improve performance, but disables support for dynamically spawning processes inside the simulation (e.g. the fork syscall).

use_cpu_pinning: Option<bool>

Pin each thread and any processes it executes to the same logical CPU Core to improve cache affinity

use_worker_spinning: Option<bool>

Each worker thread will spin in a sched_yield loop while waiting for a new task. This is ignored if not using the thread-per-core scheduler.

runahead: Option<NullableOption<Time<TimePrefix>>>

If set, overrides the automatically calculated minimum time workers may run ahead when sending events between nodes

use_dynamic_runahead: Option<bool>

Update the minimum runahead dynamically throughout the simulation.

socket_send_buffer: Option<Bytes<SiPrefixUpper>>

Initial size of the socket’s send buffer

socket_send_autotune: Option<bool>

Enable send window autotuning

socket_recv_buffer: Option<Bytes<SiPrefixUpper>>

Initial size of the socket’s receive buffer

socket_recv_autotune: Option<bool>

Enable receive window autotuning

interface_qdisc: Option<QDiscMode>

The queueing discipline to use at the network interface

host_heartbeat_log_level: Option<LogLevel>

Log level at which to print host statistics

host_heartbeat_log_info: Option<HashSet<LogInfoFlag>>

List of information to show in the host’s heartbeat message

host_heartbeat_interval: Option<NullableOption<Time<TimePrefix>>>

Amount of time between heartbeat messages for this host

strace_logging_mode: Option<StraceLoggingMode>

Log the syscalls for each process to individual “strace” files

max_unapplied_cpu_latency: Option<Time<TimePrefix>>

Max amount of execution-time latency allowed to accumulate before the clock is moved forward. Moving the clock forward is a potentially expensive operation, so larger values reduce simulation overhead, at the cost of coarser time jumps. Note also that accumulated-but-unapplied latency is discarded when a thread is blocked on a syscall.

unblocked_syscall_latency: Option<Time<TimePrefix>>

Simulated latency of an unblocked syscall. For efficiency Shadow only actually adds this latency if and when max_unapplied_cpu_latency is reached.

unblocked_vdso_latency: Option<Time<TimePrefix>>

Simulated latency of a vdso “syscall”. For efficiency Shadow only actually adds this latency if and when max_unapplied_cpu_latency is reached.

scheduler: Option<Scheduler>

The host scheduler implementation, which decides how to assign hosts to threads and threads to CPU cores

log_errors_to_tty: Option<bool>

When true, log error-level messages to stderr in addition to stdout when stdout is not a tty but stderr is.

use_new_tcp: Option<bool>

Use the rust TCP implementation

Implementations

impl ExperimentalOptions

pub fn with_defaults(self, default: Self) -> Self

Replace unset (None) values of base with values from default.