Module shadow_rs::core::support::configuration

Modules

export

Structs

CliOptions
Run real applications over simulated networks.

ConfigFileOptions
Options contained in a configuration file.

ConfigOptions
Shadow configuration options after processing command-line and configuration file options.

ExperimentalOptions

FileSource

GeneralOptions

HostDefaultOptions

HostOptions

NetworkOptions

ProcessOptions

Quantity

Enums

Compression

GraphOptions

GraphSource

InterposeMethod

LogInfoFlag

LogLevel

ProcessArgs

QDiscMode

SchedulerPolicy

Constants

END_HELP_TEXT

ONE_GBIT_SWITCH_GRAPH

Statics

EXP_HELP
Help messages used by Clap for command line arguments, combining the doc string with the Serde default.

GENERAL_HELP
Help messages used by Clap for command line arguments, combining the doc string with the Serde default.

HOST_HELP
Help messages used by Clap for command line arguments, combining the doc string with the Serde default.

NETWORK_HELP
Help messages used by Clap for command line arguments, combining the doc string with the Serde default.

Functions

default_args_empty
Helper function for serde default ProcessArgs::Str("") values.

default_data_directory

default_some_1
Helper function for serde default Some(1) values.

default_some_info
Helper function for serde default Some(LogLevel::Info) values.

default_some_nz_1
Helper function for serde default Some(1) values.

default_some_time_0
Helper function for serde default Some(0) values.

default_some_time_1
Helper function for serde default Some(0) values.

default_some_true
Helper function for serde default Some(true) values.

generate_help_strs
Generate help strings for objects in a JSON schema, including the Serde defaults if available.

parse_set_log_info_flags
Parse a string as a set of LogInfoFlag values.

parse_string_as_args
Parses a string as a list of arguments following the shell’s parsing rules. This uses g_shell_parse_argv() for parsing.

tilde_expansion