Function shadow_rs::core::work::task::export::taskref_new_bound

#[no_mangle]
pub unsafe extern "C" fn taskref_new_bound(
    host_id: HostId, 
    callback: TaskCallbackFunc, 
    object: *mut c_void, 
    argument: *mut c_void, 
    object_free: TaskObjectFreeFunc, 
    argument_free: TaskArgumentFreeFunc
) -> *mut TaskRef

Create a new reference-counted task that can only be executed on the given host. The callbacks can safely assume that they will only be called with the lock for the specified host held.

SAFETY:

object and argument must meet the requirements for HostTreePointer::new.

Given that the host lock is held when execution of a callback starts, they must not cause object or argument to be dereferenced without the host lock held. (e.g. by releasing the host lock or exfiltrating the pointers to be dereferenced by other code that might not hold the lock).

There must still be some coordination between the creator of the TaskRef and the callers of taskref_execute and taskref_drop to ensure that the callbacks don’t conflict with other accesses in the same thread (e.g. that the caller isn’t holding a Rust mutable reference to one of the pointers while the callback transforms the pointer into another Rust reference).