shadow_rs::core::work::task::export

Function taskref_new_unbound

#[unsafe(no_mangle)]
pub unsafe extern "C-unwind" fn taskref_new_unbound(
    callback: TaskCallbackFunc,
    object: *mut c_void,
    argument: *mut c_void,
    object_free: TaskObjectFreeFunc,
    argument_free: TaskArgumentFreeFunc,
) -> *mut TaskRef

Create a new reference-counted task that may be executed on any Host.

Safety

The callbacks must be safe to call with object and argument with any Host. (e.g. even if task is expected to execute on another Host, must be safe to execute or free the Task from the current Host.)

There must still be some coordination between the creator of the TaskRef and the callers of taskref_execute and taskref_drop to ensure that the callbacks don’t conflict with other accesses in the same thread (e.g. that the caller isn’t holding a Rust mutable reference to one of the pointers while the callback transforms the pointer into another Rust reference).