Struct shadow_rs::core::worker::Worker

pub struct Worker {
    worker_id: WorkerThreadID,
    active_host_info: Option<Arc<HostInfo>>,
    active_process_info: Option<ProcessInfo>,
    active_thread_info: Option<ThreadInfo>,
    cworker: *mut WorkerC,
}

Worker context, containing ‘global’ information for the current thread.

Fields
worker_id: WorkerThreadIDactive_host_info: Option<Arc<HostInfo>>active_process_info: Option<ProcessInfo>active_thread_info: Option<ThreadInfo>cworker: *mut WorkerC
Implementations

impl Worker

pub unsafe fn new_for_this_thread(
cworker: *mut WorkerC,
worker_id: WorkerThreadID
)

pub fn with_active_host_info<F, R>(f: F) -> Option<R> where
F: FnOnce(&Arc<HostInfo>) -> R,

Run f with a reference to the current Host, or return None if there is no current Host.

pub fn set_active_host(host: &Host)

Set the currently-active Host.

pub fn clear_active_host()

Clear the currently-active Host.

pub fn set_active_process(process: &Process)

Set the currently-active Process.

pub fn clear_active_process()

Clear the currently-active Process.

pub fn set_active_thread(thread: &CThread)

Set the currently-active Thread.

pub fn clear_active_thread()

Clear the currently-active Thread.

pub fn is_alive() -> bool

Whether currently running on a live Worker.

pub fn current_time() -> Option<SimulationTime>

Current simulation time, or None if not running on a live Worker.

pub fn thread_id() -> Option<WorkerThreadID>

ID of this thread’s Worker, if any.

pub fn active_process_native_pid() -> Option<Pid>

pub fn active_process_id() -> Option<ProcessId>

pub fn active_thread_native_tid() -> Option<Pid>