Struct shadow_rs::core::worker::Worker

pub struct Worker {
    worker_id: WorkerThreadID,
    active_host_info: Option<Arc<HostInfo>>,
    active_process_info: Option<ProcessInfo>,
    active_thread_info: Option<ThreadInfo>,
    clock: Clock,
    bootstrap_end_time: SimulationTime,
    syscall_counter: Counter,
    object_alloc_counter: Counter,
    object_dealloc_counter: Counter,
    worker_pool: *mut u8,
}

Worker context, containing ‘global’ information for the current thread.

Fields
worker_id: WorkerThreadIDactive_host_info: Option<Arc<HostInfo>>active_process_info: Option<ProcessInfo>active_thread_info: Option<ThreadInfo>clock: Clockbootstrap_end_time: SimulationTimesyscall_counter: Counterobject_alloc_counter: Counterobject_dealloc_counter: Counterworker_pool: *mut u8
Implementations

impl Worker

pub unsafe fn new_for_this_thread(
worker_pool: *mut u8,
worker_id: WorkerThreadID,
bootstrap_end_time: SimulationTime
)

pub fn with_active_host_info<F, R>(f: F) -> Option<R> where
F: FnOnce(&Arc<HostInfo>) -> R,

Run f with a reference to the current Host, or return None if there is no current Host.

pub fn set_active_host(host: &Host)

Set the currently-active Host.

pub fn clear_active_host()

Clear the currently-active Host.

pub fn set_active_process(process: &Process)

Set the currently-active Process.

pub fn clear_active_process()

Clear the currently-active Process.

pub fn set_active_thread(thread: &CThread)

Set the currently-active Thread.

pub fn clear_active_thread()

Clear the currently-active Thread.

pub fn is_alive() -> bool

Whether currently running on a live Worker.

pub fn thread_id() -> Option<WorkerThreadID>

ID of this thread’s Worker, if any.

pub fn active_process_native_pid() -> Option<Pid>

pub fn active_process_id() -> Option<ProcessId>

pub fn active_thread_native_tid() -> Option<Pid>

fn set_round_end_time(t: SimulationTime)

fn round_end_time() -> Option<SimulationTime>

fn set_current_time(t: SimulationTime)

fn clear_current_time()

pub fn current_time() -> Option<SimulationTime>

fn set_last_event_time(t: SimulationTime)

fn with<F, O>(f: F) -> Option<O> where
F: FnOnce(&Worker) -> O,

fn with_mut<F, O>(f: F) -> Option<O> where
F: FnOnce(&mut Worker) -> O,