Struct shadow_rs::core::worker::Worker

pub struct Worker { /* private fields */ }

Worker context, containing ‘global’ information for the current thread.

Implementations

impl Worker

pub unsafe fn new_for_this_thread(worker_id: WorkerThreadID)

pub fn with_active_host_info<F, R>(f: F) -> Option<R> where
F: FnOnce(&Arc<HostInfo>) -> R,

Run f with a reference to the current Host, or return None if there is no current Host.

pub fn set_active_host(host: &Host)

Set the currently-active Host.

pub fn clear_active_host()

Clear the currently-active Host.

pub fn set_active_process(process: &Process)

Set the currently-active Process.

pub fn clear_active_process()

Clear the currently-active Process.

pub fn set_active_thread(thread: &CThread)

Set the currently-active Thread.

pub fn clear_active_thread()

Clear the currently-active Thread.

pub fn is_alive() -> bool

Whether currently running on a live Worker.

pub fn thread_id() -> Option<WorkerThreadID>

ID of this thread’s Worker, if any.

pub fn active_process_native_pid() -> Option<Pid>

pub fn active_process_id() -> Option<ProcessId>

pub fn active_thread_native_tid() -> Option<Pid>

pub fn set_round_end_time(t: EmulatedTime)

pub fn set_current_time(t: EmulatedTime)

pub fn clear_current_time()

pub fn current_time() -> Option<EmulatedTime>

pub fn update_lowest_used_latency(t: SimulationTime)

pub fn reset_next_event_time()

pub fn get_next_event_time() -> Option<EmulatedTime>

pub fn update_next_event_time(t: EmulatedTime)

pub fn increment_object_alloc_counter(s: &str)

pub fn increment_object_dealloc_counter(s: &str)

pub fn add_to_global_alloc_counters()

pub fn worker_add_syscall_counts(syscall_counts: &Counter)