Struct shadow_rs::core::worker::Worker

pub struct Worker { /* private fields */ }

Worker context, containing ‘global’ information for the current thread.

Implementations

impl Worker

pub fn new_for_this_thread(worker_id: WorkerThreadID)

pub fn with_active_host<F, R>(f: F) -> Option<R>
where F: FnOnce(&Host) -> R,

Run f with a reference to the current Host, or return None if there is no current Host.

pub fn with_active_process_rc<F, R>(f: F) -> Option<R>
where F: FnOnce(&RootedRc<RootedRefCell<Process>>) -> R,

Run f with a reference to the current RootedRc<RootedRefCell<Process>>, or return None if there isn’t one.

Prefer to pass Process explicitly where feasible. e.g. see ProcessContext.

pub fn with_active_process<F, R>(f: F) -> Option<R>
where F: FnOnce(&Process) -> R,

Run f with a reference to the current Process, or return None if there isn’t one.

Prefer to pass Process explicitly where feasible. e.g. see ProcessContext.

pub fn with_active_thread<F, R>(f: F) -> Option<R>
where F: FnOnce(&Thread) -> R,

Run f with a reference to the current Thread, or return None if there isn’t one.

Prefer to pass Thread explicitly where feasible. e.g. see ThreadContext.

pub fn with_dns<F, R>(f: F) -> R
where F: FnOnce(&DNS) -> R,

Run f with a reference to the global DNS.

Panics if the Worker or its DNS hasn’t yet been initialized.

pub fn set_active_host(host: Box<Host>)

Set the currently-active Host.

pub fn take_active_host() -> Box<Host>

Clear the currently-active Host.

pub fn set_active_process(process: &RootedRc<RootedRefCell<Process>>)

Set the currently-active Process.

pub fn clear_active_process()

Clear the currently-active Process.

pub fn set_active_thread(thread: &RootedRc<RootedRefCell<Thread>>)

Set the currently-active Thread.

pub fn clear_active_thread()

Clear the currently-active Thread.

pub fn is_alive() -> bool

Whether currently running on a live Worker.

pub fn worker_id() -> Option<WorkerThreadID>

ID of this thread’s Worker, if any.

pub fn active_process_native_pid() -> Option<Pid>

pub fn active_process_id() -> Option<ProcessId>

pub fn active_thread_id() -> Option<ThreadId>

pub fn active_thread_native_tid() -> Option<Pid>

pub fn set_round_end_time(t: EmulatedTime)

pub fn max_event_runahead_time(host: &Host) -> EmulatedTime

Maximum time that the current event may run ahead to.

pub fn set_current_time(t: EmulatedTime)

pub fn clear_current_time()

pub fn current_time() -> Option<EmulatedTime>

pub fn update_lowest_used_latency(t: SimulationTime)

pub fn reset_next_event_time()

pub fn get_next_event_time() -> Option<EmulatedTime>

pub fn update_next_event_time(t: EmulatedTime)

pub unsafe fn send_packet(src_host: &Host, packet: *mut Packet)

Safety

packet must be valid and not accessed by another thread while this function is running.

pub fn increment_object_alloc_counter(s: &str)

pub fn increment_object_dealloc_counter(s: &str)

pub fn add_syscall_counts(syscall_counts: &Counter)

pub fn add_to_global_sim_stats()

pub fn is_routable(src: IpAddr, dst: IpAddr) -> bool

pub fn increment_plugin_error_count()

pub fn is_bootstrapping() -> bool

Shadow allows configuration of a “bootstrapping” interval, during which hosts’ network activity does not consume bandwidth. Returns true if we are still within this preliminary interval, or false otherwise.

pub fn resolve_name_to_ip(name: &CStr) -> Option<Ipv4Addr>