Module shadow_rs::cshadow

Re-exports

pub use self::_LogLevel as LogLevel;

pub use self::_Status as Status;

pub use self::_LogInfoFlags as LogInfoFlags;

pub use self::_ProtocolType as ProtocolType;

pub use self::_TriggerType as TriggerType;

Structs

SysCallReturnBlocked

SysCallReturnDone

_Address

_CPU

_DNS

_Epoll

_Futex

_FutexTable

_GTimer

_Host

_HostParameters

_NetworkInterface

_Packet

_PacketTCPHeader

_PluginPhysicalPtr

_PluginVirtualPtr

_Process

_Router

_ShMemBlock

_ShimHostProtectedSharedMem

_ShimShmemHost

_ShimShmemProcess

_ShimShmemThread

_StatusListener

_SysCallArgs

_SysCallCondition

_SysCallHandler

_SysCallReturn

_Thread

_Tracker

_Transport

_TransportFunctionTable

_Trigger

_Tsc

__sigset_t

siginfo_t

siginfo_t__bindgen_ty_1__bindgen_ty_1

siginfo_t__bindgen_ty_1__bindgen_ty_2

siginfo_t__bindgen_ty_1__bindgen_ty_3

siginfo_t__bindgen_ty_1__bindgen_ty_4

siginfo_t__bindgen_ty_1__bindgen_ty_5

siginfo_t__bindgen_ty_1__bindgen_ty_5__bindgen_ty_1__bindgen_ty_1

siginfo_t__bindgen_ty_1__bindgen_ty_6

siginfo_t__bindgen_ty_1__bindgen_ty_7

Constants

CONFIG_PIPE_BUFFER_SIZE

ProtocolTCPFlags_PTCP_ACK

ProtocolTCPFlags_PTCP_DUPACK

ProtocolTCPFlags_PTCP_FIN

ProtocolTCPFlags_PTCP_NONE

ProtocolTCPFlags_PTCP_RST

ProtocolTCPFlags_PTCP_SACK

ProtocolTCPFlags_PTCP_SYN

SHADOW_SOMAXCONN

SYSCALL_IO_BUFSIZE

SysCallReturnState_SYSCALL_BLOCK

SysCallReturnState_SYSCALL_DONE

SysCallReturnState_SYSCALL_NATIVE

_LogInfoFlags_LOG_INFO_FLAGS_NODE

_LogInfoFlags_LOG_INFO_FLAGS_NONE

_LogInfoFlags_LOG_INFO_FLAGS_RAM

_LogInfoFlags_LOG_INFO_FLAGS_SOCKET

_LogLevel_LOGLEVEL_DEBUG

_LogLevel_LOGLEVEL_ERROR

_LogLevel_LOGLEVEL_INFO

_LogLevel_LOGLEVEL_TRACE

_LogLevel_LOGLEVEL_UNSET

_LogLevel_LOGLEVEL_WARNING

_ProtocolType_PLOCAL

_ProtocolType_PNONE

_ProtocolType_PTCP

_ProtocolType_PUDP

_Status_STATUS_FILE_ACTIVE

_Status_STATUS_FILE_CLOSED

_Status_STATUS_FILE_READABLE

_Status_STATUS_FILE_WRITABLE

_Status_STATUS_FUTEX_WAKEUP

_Status_STATUS_NONE

_Status_STATUS_SOCKET_ALLOWING_CONNECT

_TriggerType_TRIGGER_DESCRIPTOR

_TriggerType_TRIGGER_FILE

_TriggerType_TRIGGER_FUTEX

_TriggerType_TRIGGER_NONE

Functions

affinity_getGoodWorkerAffinity⚠

affinity_initPlatformInfo⚠

affinity_setProcessAffinity⚠

dns_deregister⚠

dns_free⚠

dns_getHostsFilePath⚠

dns_new⚠

dns_register⚠

dns_resolveIPToAddress⚠

dns_resolveNameToAddress⚠

g_quark_from_string⚠

host_addApplication⚠

host_autotuneReceiveBuffer⚠

host_autotuneSendBuffer⚠

host_boot⚠

host_continueExecutionTimer⚠

host_doesInterfaceExist⚠

host_execute⚠

host_freeAllApplications⚠

host_getAbstractUnixNamespace⚠

host_getCPU⚠

host_getConfiguredRecvBufSize⚠

host_getConfiguredSendBufSize⚠

host_getDataPath⚠

host_getDefaultAddress⚠

host_getDefaultIP⚠

host_getFutexTable⚠

host_getID⚠

host_getLogLevel⚠

host_getName⚠

host_getNativeTID⚠

host_getNewEventID⚠

host_getNewPacketID⚠

host_getNewProcessID⚠

host_getNextDeterministicSequenceValue⚠

host_getNextPacketPriority⚠

host_getProcess⚠

host_getRandom⚠

host_getRandomFreePort⚠

host_getSharedMem⚠

host_getShimShmemLock⚠

host_getThread⚠

host_getTracker⚠

host_getTsc⚠

host_getUpstreamRouter⚠

host_get_bw_down_kiBps⚠

host_get_bw_up_kiBps⚠

host_isInterfaceAvailable⚠

host_lock⚠

host_lockShimShmemLock⚠

host_new⚠

host_nextEventTime⚠

host_pushLocalEvent⚠

host_ref⚠

host_returnHandleHack⚠

host_scheduleTaskAtEmulatedTime⚠

host_scheduleTaskWithDelay⚠

host_setup⚠

host_shutdown⚠

host_stopExecutionTimer⚠

host_unlock⚠

host_unlockShimShmemLock⚠

host_unref⚠

legacyfile_close⚠

legacyfile_ref⚠

legacyfile_shutdownHelper⚠

legacyfile_unref⚠

logger_get_global_start_time_micros⚠

main_checkGlibVersion⚠

main_logBuildInfo⚠

main_printBuildInfo⚠

main_sidechannelMitigationsEnabled⚠

packet_copyPayloadShadow⚠

packet_getDestinationIP⚠

packet_getDestinationPort⚠

packet_getHeaderSize⚠

packet_getPayloadSize⚠

packet_getProtocol⚠

packet_getSourceIP⚠

packet_getSourcePort⚠

packet_getTCPHeader⚠

packet_getTotalSize⚠

process_addThread⚠

process_continue⚠

process_detachPlugin⚠

process_findNativeTID⚠

process_flushPtrs⚠

process_freePtrsWithoutFlushing⚠

process_getDescriptorTable⚠

process_getHostId⚠

process_getMemoryManager⚠

process_getMutablePtr⚠

process_getName⚠

process_getNativePid⚠

process_getPhysicalAddress⚠

process_getPluginName⚠

process_getProcessID⚠

process_getReadablePtr⚠

process_getReadableString⚠

process_getRealtimeTimer⚠

process_getSharedMem⚠

process_getStraceFd⚠

process_getThread⚠

process_getWorkingDir⚠

process_getWriteablePtr⚠

process_hasStarted⚠

process_isRunning⚠

process_markAsExiting⚠

process_new⚠

process_parseArgStr⚠

process_parseArgStrFree⚠

process_readPtr⚠

process_readString⚠

process_ref⚠

process_schedule⚠

process_setMemoryManager⚠

process_signal⚠

process_stop⚠

process_straceLoggingMode⚠

process_unref⚠

process_writePtr⚠

return_code_for_signal⚠

runConfigHandlers⚠

rustlogger_new⚠

scanRpathForLib⚠

scheduler_addHost⚠

scheduler_awaitNextRound⚠

scheduler_continueNextRound⚠

scheduler_finish⚠

scheduler_free⚠

scheduler_new⚠

scheduler_shutdown⚠

scheduler_start⚠

statuslistener_onStatusChanged⚠

statuslistener_ref⚠

statuslistener_unref⚠

syscallcondition_getActiveFile⚠

syscallcondition_new⚠

syscallcondition_setActiveFile⚠

syscallcondition_unref⚠

syscallhandler_accept⚠

syscallhandler_accept4⚠

syscallhandler_bind⚠

syscallhandler_connect⚠

syscallhandler_exit_group⚠

syscallhandler_fcntl⚠

syscallhandler_getpeername⚠

syscallhandler_getpid⚠

syscallhandler_getppid⚠

syscallhandler_getsockname⚠

syscallhandler_getsockopt⚠

syscallhandler_ioctl⚠

syscallhandler_listen⚠

syscallhandler_pread64⚠

syscallhandler_pwrite64⚠

syscallhandler_read⚠

syscallhandler_recvfrom⚠

syscallhandler_sendto⚠

syscallhandler_set_tid_address⚠

syscallhandler_setsockopt⚠

syscallhandler_shutdown⚠

syscallhandler_socket⚠

syscallhandler_socketpair⚠

syscallhandler_uname⚠

syscallhandler_write⚠

thread_clearSysCallCondition⚠

thread_clone⚠

thread_getHost⚠

thread_getHostId⚠

thread_getID⚠

thread_getIPCBlock⚠

thread_getNativePid⚠

thread_getNativeTid⚠

thread_getProcess⚠

thread_getProcessId⚠

thread_getReturnCode⚠

thread_getShMBlock⚠

thread_getSignalSet⚠

thread_getSysCallCondition⚠

thread_getSysCallHandler⚠

thread_getTidAddress⚠

thread_handleProcessExit⚠

thread_isLeader⚠

thread_isRunning⚠

thread_nativeSyscall⚠

thread_new⚠

thread_ref⚠

thread_resume⚠

thread_run⚠

thread_setTidAddress⚠

thread_unblockedSignalPending⚠

thread_unref⚠

worker_clearCurrentTime⚠

worker_getAffinity⚠

worker_getChildPidWatcher⚠

worker_getConfig⚠

worker_getCurrentEmulatedTime⚠

worker_getCurrentSimulationTime⚠

worker_getDNS⚠

worker_getLatencyForAddresses⚠

worker_getNodeBandwidthDownKiBps⚠

worker_getNodeBandwidthUpKiBps⚠

worker_getReliabilityForAddresses⚠

worker_incrementPacketCount⚠

worker_incrementPluginError⚠

worker_isAlive⚠

worker_isBootstrapActive⚠

worker_isFiltered⚠

worker_isRoutable⚠

worker_maxEventRunaheadTime⚠

worker_resolveIPToAddress⚠

worker_resolveNameToAddress⚠

worker_scheduleTaskAtEmulatedTime⚠

worker_scheduleTaskWithDelay⚠

worker_sendPacket⚠

worker_setCurrentEmulatedTime⚠

worker_setMinEventTimeNextRound⚠

worker_setRoundEndTime⚠

workerpool_updateMinHostRunahead⚠

Type Definitions

Address

An Address structure holds information used to identify nodes, allowing for easy extraction of both integer and string forms of an IP address as well as the string hostname associated with the IP. Address is an opaque structure and should only be accessed using the functions in this class.

CPU

DNS

EmulatedTime

Emulation time in nanoseconds. Allows for a consistent representation of time throughput the simulator. Emulation time is the simulation time plus the EMULATION_TIME_OFFSET. This type allows us to explicitly distinguish each type of time in the code.,

Epoll

Futex

FutexTable

GList

GQuark

GTimer

Host

HostId

HostParameters

LegacyFile

LegacyFileCleanupFunc

LegacyFileCloseFunc

LegacyFileFreeFunc

NetworkInterface

Packet

PacketTCPHeader

PluginPhysicalPtr

PluginPtr

PluginVirtualPtr

Process

ProtocolTCPFlags

Router

Scheduler

ShMemBlock

ShimShmemHost

ShimShmemHostLock

ShimShmemProcess

ShimShmemThread

SimulationTime

Simulation time in nanoseconds. Allows for a consistent representation of time throughput the simulator.

StatusListener

SysCallArgs

SysCallCondition

SysCallHandler

SysCallReg

SysCallReturn

SysCallReturnState

Thread

Tracker

Transport

TransportFunctionTable

TransportReceiveFunc

TransportSendFunc

Trigger

TriggerObject

Tsc

WorkerPool

_LogInfoFlags

_LogLevel

_ProtocolType

_Status

_TriggerType

__clock_t

__int32_t

__int64_t

__pid_t

__sigval_t

__ssize_t

__uid_t

__uint16_t

__uint32_t

__uint64_t

gboolean

gchar

gdouble

gint

gpointer

gsize

gssize

guint

guint32

guint64

gulong

in_addr_t

in_port_t

pid_t

sa_family_t

sigset_t

size_t

ssize_t

Unions

SysCallReturnBody

_SysCallReg

_TriggerObject

siginfo_t__bindgen_ty_1

siginfo_t__bindgen_ty_1__bindgen_ty_5__bindgen_ty_1

sigval