Module shadow_rs::cshadow

cbindgen:ignore

Re-exports

pub use self::_LogLevel as LogLevel;

pub use self::_LegacyFileType as LegacyFileType;

pub use self::_PacketDeliveryStatusFlags as PacketDeliveryStatusFlags;

pub use self::_StatusListenerFilter as StatusListenerFilter;

pub use self::_ProtocolType as ProtocolType;

pub use self::_LogInfoFlags as LogInfoFlags;

pub use self::_CompatSocketTypes as CompatSocketTypes;

pub use self::_FileType as FileType;

pub use self::_TriggerType as TriggerType;

Structs

CompatUntypedForeignPtr

SysCallCondition

TCPCongHooks_

TCPCong_

_Address

_CompatSocket

_DNS

_Epoll

_Futex

_GList

_LegacyFileFunctionTable

_NetworkInterface

_Packet

_PacketTCPHeader

_RegularFile

_StatusListener

_TCP

_Tracker

_Trigger

__BindgenBitfieldUnit

__fsid_t

iovec

linux___sifields__bindgen_ty_1

linux___sifields__bindgen_ty_2

linux___sifields__bindgen_ty_3

linux___sifields__bindgen_ty_4

linux___sifields__bindgen_ty_5

linux___sifields__bindgen_ty_5__bindgen_ty_1__bindgen_ty_1

linux___sifields__bindgen_ty_5__bindgen_ty_1__bindgen_ty_2

linux___sifields__bindgen_ty_5__bindgen_ty_1__bindgen_ty_3

linux___sifields__bindgen_ty_6

linux___sifields__bindgen_ty_7

linux_dirent

linux_dirent64

linux_siginfo

linux_siginfo__bindgen_ty_1__bindgen_ty_1

pollfd

stat

statfs

statx

statx_timestamp

tcp_info

timespec

timeval

Constants

AFFINITY_UNINIT

CONFIG_HEADER_SIZE_TCP

CONFIG_MTU

CONFIG_PIPE_BUFFER_SIZE

GLIB_MAJOR_VERSION

GLIB_MICRO_VERSION

GLIB_MINOR_VERSION

ProtocolTCPFlags_PTCP_ACK

ProtocolTCPFlags_PTCP_DUPACK

ProtocolTCPFlags_PTCP_FIN

ProtocolTCPFlags_PTCP_NONE

ProtocolTCPFlags_PTCP_RST

ProtocolTCPFlags_PTCP_SACK

ProtocolTCPFlags_PTCP_SYN

ProtocolUDPFlags_PUDP_NONE

SHADOW_SOMAXCONN

SYSCALL_IO_BUFSIZE

ShadowSyscallNum_SYS_shadow_hostname_to_addr_ipv4

ShadowSyscallNum_SYS_shadow_init_memory_manager

ShadowSyscallNum_SYS_shadow_max

ShadowSyscallNum_SYS_shadow_min

ShadowSyscallNum_SYS_shadow_yield

_CompatSocketTypes_CST_INET_SOCKET

_CompatSocketTypes_CST_LEGACY_SOCKET

_CompatSocketTypes_CST_NONE

_FileType_FILE_TYPE_HOSTS

_FileType_FILE_TYPE_IN_MEMORY

_FileType_FILE_TYPE_LOCALTIME

_FileType_FILE_TYPE_NOTSET

_FileType_FILE_TYPE_RANDOM

_FileType_FILE_TYPE_REGULAR

_LegacyFileType_DT_EPOLL

_LegacyFileType_DT_FILE

_LegacyFileType_DT_NONE

_LegacyFileType_DT_TCPSOCKET

_LogInfoFlags_LOG_INFO_FLAGS_NODE

_LogInfoFlags_LOG_INFO_FLAGS_NONE

_LogInfoFlags_LOG_INFO_FLAGS_RAM

_LogInfoFlags_LOG_INFO_FLAGS_SOCKET

_LogLevel_LOGLEVEL_DEBUG

_LogLevel_LOGLEVEL_ERROR

_LogLevel_LOGLEVEL_INFO

_LogLevel_LOGLEVEL_TRACE

_LogLevel_LOGLEVEL_UNSET

_LogLevel_LOGLEVEL_WARNING

_PacketDeliveryStatusFlags_PDS_DESTROYED

_PacketDeliveryStatusFlags_PDS_INET_DROPPED

_PacketDeliveryStatusFlags_PDS_INET_SENT

_PacketDeliveryStatusFlags_PDS_NONE

_PacketDeliveryStatusFlags_PDS_RCV_INTERFACE_DROPPED

_PacketDeliveryStatusFlags_PDS_RCV_INTERFACE_RECEIVED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_BUFFERED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_DELIVERED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_DROPPED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_PROCESSED

_PacketDeliveryStatusFlags_PDS_RCV_TCP_ENQUEUE_UNORDERED

_PacketDeliveryStatusFlags_PDS_RELAY_CACHED

_PacketDeliveryStatusFlags_PDS_RELAY_FORWARDED

_PacketDeliveryStatusFlags_PDS_ROUTER_DEQUEUED

_PacketDeliveryStatusFlags_PDS_ROUTER_DROPPED

_PacketDeliveryStatusFlags_PDS_ROUTER_ENQUEUED

_PacketDeliveryStatusFlags_PDS_SND_CREATED

_PacketDeliveryStatusFlags_PDS_SND_INTERFACE_SENT

_PacketDeliveryStatusFlags_PDS_SND_SOCKET_BUFFERED

_PacketDeliveryStatusFlags_PDS_SND_TCP_DEQUEUE_RETRANSMIT

_PacketDeliveryStatusFlags_PDS_SND_TCP_ENQUEUE_RETRANSMIT

_PacketDeliveryStatusFlags_PDS_SND_TCP_ENQUEUE_THROTTLED

_PacketDeliveryStatusFlags_PDS_SND_TCP_RETRANSMITTED

_ProtocolType_PMOCK

_ProtocolType_PNONE

_ProtocolType_PTCP

_ProtocolType_PUDP

_StatusListenerFilter_SLF_ALWAYS

_StatusListenerFilter_SLF_NEVER

_StatusListenerFilter_SLF_OFF_TO_ON

_StatusListenerFilter_SLF_ON_TO_OFF

_TriggerType_TRIGGER_CHILD

_TriggerType_TRIGGER_DESCRIPTOR

_TriggerType_TRIGGER_FILE

_TriggerType_TRIGGER_FUTEX

_TriggerType_TRIGGER_NONE

Statics

SHADOW_FLAG_MASK

TCP_CONG_RENO_NAME

glib_major_version

glib_micro_version

glib_minor_version

Functions

_syscallhandler_pollHelper⚠

_syscallhandler_readHelper⚠

_syscallhandler_validateLegacyFile⚠

_syscallhandler_writeHelper⚠

address_getID⚠

address_ipToNewString⚠

Turns the IPv4 address into a newly allocated string that should be freed by the caller. Address must be in network byte order.

address_isEqual⚠

Checks if the given addresses are equal. This function is NULL safe, so so either or both addresses may be NULL. @param a a valid, non-NULL Address structure previously created with address_new() @param b a valid, non-NULL Address structure previously created with address_new() @return TRUE if both addresses are NULL or if the IP associated with both addresses are equal, FALSE otherwise

address_isLocal⚠

address_new⚠

Create a new Address structure with the given IP and Hostname. The hostname is copied and stored internally, so the caller retains its ownership.

address_ref⚠

address_stringToIP⚠

address_toHostIP⚠

Retrieve the host-order integer version of this address @param address a valid, non-NULL Address structure previously created with address_new() @return the host-order IP

address_toHostIPString⚠

Retrieves the dot-and-decimal string representation of the host-order version of this address. The caller does not own and should not modify or free the string. @param address a valid, non-NULL Address structure previously created with address_new() @return an address pointing to the internal memory holding the string

address_toHostName⚠

Retrieves the hostname of this address. The caller does not own and should not modify or free the string. @param address a valid, non-NULL Address structure previously created with address_new() @return an address pointing to the internal memory holding the string

address_toNetworkIP⚠

Retrieve the network-order integer version of this address @param address a valid, non-NULL Address structure previously created with address_new() @return the network-order IP

address_toString⚠

address_unref⚠

affinity_getGoodWorkerAffinity⚠

affinity_initPlatformInfo⚠

affinity_setProcessAffinity⚠

compatsocket_fromInetSocket⚠

compatsocket_fromLegacySocket⚠

compatsocket_getCanonicalHandle⚠

dns_deregister⚠

dns_free⚠

dns_getHostsFilePath⚠

dns_new⚠

dns_register⚠

dns_resolveIPToAddress⚠

dns_resolveNameToAddress⚠

epoll_new⚠

futex_addListener⚠

futex_getAddress⚠

futex_getListenerCount⚠

futex_new⚠

futex_ref⚠

futex_removeListener⚠

futex_unref⚠

futex_unref_func⚠

futex_wake⚠

g_list_append⚠

g_list_free⚠

glib_check_version⚠

legacyfile_addFlags⚠

legacyfile_addListener⚠

legacyfile_adjustStatus⚠

legacyfile_clear⚠

legacyfile_close⚠

legacyfile_getEventSource⚠

legacyfile_getFlags⚠

legacyfile_getStatus⚠

legacyfile_getType⚠

legacyfile_init⚠

legacyfile_ref⚠

legacyfile_refWeak⚠

legacyfile_removeFlags⚠

legacyfile_removeListener⚠

legacyfile_setFlags⚠

legacyfile_supportsSaRestart⚠

legacyfile_unref⚠

legacyfile_unrefWeak⚠

legacysocket_addToInputBuffer⚠

legacysocket_addToOutputBuffer⚠

legacysocket_connectToPeer⚠

legacysocket_dropPacket⚠

legacysocket_getInputBufferLength⚠

legacysocket_getInputBufferSize⚠

legacysocket_getInputBufferSpace⚠

legacysocket_getOutputBufferLength⚠

legacysocket_getOutputBufferSize⚠

legacysocket_getOutputBufferSpace⚠

legacysocket_getPeerName⚠

legacysocket_getProtocol⚠

legacysocket_getSocketName⚠

legacysocket_getUnixPath⚠

legacysocket_isBound⚠

legacysocket_isFamilySupported⚠

legacysocket_isUnix⚠

legacysocket_peekNextInPacket⚠

legacysocket_peekNextOutPacket⚠

legacysocket_pullOutPacket⚠

legacysocket_pushInPacket⚠

legacysocket_receiveUserData⚠

legacysocket_removeFromInputBuffer⚠

legacysocket_removeFromOutputBuffer⚠

legacysocket_sendUserData⚠

legacysocket_setInputBufferSize⚠

legacysocket_setOutputBufferSize⚠

legacysocket_setPeerName⚠

legacysocket_setSocketName⚠

legacysocket_setUnix⚠

legacysocket_setUnixPath⚠

logger_get_global_start_time_micros⚠

networkinterface_associate⚠

networkinterface_disassociate⚠

networkinterface_free⚠

networkinterface_isAssociated⚠

networkinterface_new⚠

networkinterface_pop⚠

networkinterface_push⚠

networkinterface_removeAllSockets⚠

networkinterface_wantsSend⚠

packet_addDeliveryStatus⚠

packet_compareTCPSequence⚠

packet_copy⚠

packet_copyPayload⚠

packet_copyPayloadShadow⚠

packet_copyPayloadWithMemoryManager⚠

packet_copyTCPSelectiveACKs⚠

packet_getDeliveryStatus⚠

packet_getDestinationIP⚠

packet_getDestinationPort⚠

packet_getHeaderSize⚠

packet_getPayloadSize⚠

packet_getPriority⚠

packet_getProtocol⚠

packet_getSourceIP⚠

packet_getSourcePort⚠

packet_getTCPHeader⚠

packet_getTotalSize⚠

packet_new⚠

packet_new_inner⚠

packet_ref⚠

packet_setMock⚠

packet_setPayload⚠

packet_setPayloadFromShadow⚠

packet_setPayloadWithMemoryManager⚠

packet_setPriority⚠

packet_setTCP⚠

packet_setUDP⚠

packet_toString⚠

packet_unref⚠

packet_updateTCP⚠

process_parseArgStr⚠

process_parseArgStrFree⚠

process_vm_readv⚠

process_vm_writev⚠

regularfile_faccessat⚠

regularfile_fadvise⚠

regularfile_fallocate⚠

regularfile_fchmod⚠

regularfile_fchmodat⚠

regularfile_fchown⚠

regularfile_fchownat⚠

regularfile_fcntl⚠

regularfile_fgetxattr⚠

regularfile_flistxattr⚠

regularfile_flock⚠

regularfile_fremovexattr⚠

regularfile_fsetxattr⚠

regularfile_fstat⚠

regularfile_fstatat⚠

regularfile_fstatfs⚠

regularfile_fsync⚠

regularfile_ftruncate⚠

regularfile_futimesat⚠

regularfile_getFlagsAtOpen⚠

regularfile_getModeAtOpen⚠

regularfile_getOSBackedFD⚠

regularfile_getShadowFlags⚠

regularfile_getType⚠

regularfile_getdents⚠

regularfile_getdents64⚠

regularfile_ioctl⚠

regularfile_linkat⚠

regularfile_lseek⚠

regularfile_mkdirat⚠

regularfile_mknodat⚠

regularfile_new⚠

regularfile_open⚠

regularfile_openat⚠

regularfile_poll⚠

regularfile_pread⚠

regularfile_preadv⚠

regularfile_preadv2⚠

regularfile_pwrite⚠

regularfile_pwritev⚠

regularfile_pwritev2⚠

regularfile_read⚠

regularfile_readahead⚠

regularfile_readlinkat⚠

regularfile_renameat2⚠

regularfile_statx⚠

regularfile_symlinkat⚠

regularfile_sync_range⚠

regularfile_unlinkat⚠

regularfile_utimensat⚠

regularfile_write⚠

scanRpathForLib⚠

status_listener_compare⚠

statuslistener_new⚠

statuslistener_onStatusChanged⚠

statuslistener_ref⚠

statuslistener_setMonitorStatus⚠

statuslistener_unref⚠

syscallcondition_cancel⚠

syscallcondition_getActiveFile⚠

syscallcondition_getTimeout⚠

syscallcondition_new⚠

syscallcondition_newWithAbsTimeout⚠

syscallcondition_ref⚠

syscallcondition_setActiveFile⚠

syscallcondition_setTimeout⚠

syscallcondition_unref⚠

syscallcondition_waitNonblock⚠

syscallcondition_wakeupForSignal⚠

syscallhandler_creat⚠

syscallhandler_faccessat⚠

syscallhandler_fadvise64⚠

syscallhandler_fallocate⚠

syscallhandler_fchmod⚠

syscallhandler_fchmodat⚠

syscallhandler_fchown⚠

syscallhandler_fchownat⚠

syscallhandler_fcntl⚠

syscallhandler_fdatasync⚠

syscallhandler_fgetxattr⚠

syscallhandler_flistxattr⚠

syscallhandler_flock⚠

syscallhandler_fremovexattr⚠

syscallhandler_fsetxattr⚠

syscallhandler_fstat⚠

syscallhandler_fstatfs⚠

syscallhandler_fsync⚠

syscallhandler_ftruncate⚠

syscallhandler_futex⚠

syscallhandler_futimesat⚠

syscallhandler_getdents⚠

syscallhandler_getdents64⚠

syscallhandler_ioctl⚠

syscallhandler_kill⚠

syscallhandler_linkat⚠

syscallhandler_lseek⚠

syscallhandler_mkdirat⚠

syscallhandler_mknodat⚠

syscallhandler_newfstatat⚠

syscallhandler_open⚠

syscallhandler_openat⚠

syscallhandler_poll⚠

syscallhandler_ppoll⚠

syscallhandler_pread64⚠

syscallhandler_preadv⚠

syscallhandler_preadv2⚠

syscallhandler_pselect6⚠

syscallhandler_pwrite64⚠

syscallhandler_pwritev⚠

syscallhandler_pwritev2⚠

syscallhandler_read⚠

syscallhandler_readahead⚠

syscallhandler_readlinkat⚠

syscallhandler_readv⚠

syscallhandler_renameat⚠

syscallhandler_renameat2⚠

syscallhandler_rt_sigaction⚠

syscallhandler_rt_sigprocmask⚠

syscallhandler_select⚠

syscallhandler_sigaltstack⚠

syscallhandler_statx⚠

syscallhandler_symlinkat⚠

syscallhandler_sync_file_range⚠

syscallhandler_syncfs⚠

syscallhandler_tgkill⚠

syscallhandler_tkill⚠

syscallhandler_unlinkat⚠

syscallhandler_utimensat⚠

syscallhandler_write⚠

syscallhandler_writev⚠

tcp_acceptServerPeer⚠

tcp_clearAllChildrenIfServer⚠

tcp_cong⚠

tcp_cong_reno_init⚠

tcp_disableReceiveBufferAutotuning⚠

tcp_disableSendBufferAutotuning⚠

tcp_enterServerMode⚠

tcp_getConnectionError⚠

tcp_getInfo⚠

tcp_getInputBufferLength⚠

tcp_getNotSentBytes⚠

tcp_getOutputBufferLength⚠

tcp_isListeningAllowed⚠

tcp_isValidListener⚠

tcp_networkInterfaceIsAboutToSendPacket⚠

tcp_new⚠

tcp_receiveUserData⚠

tcp_sendUserData⚠

tcp_setRustSocket⚠

tcp_shutdown⚠

tcp_updateServerBacklog⚠

tcpcong_nameStr⚠

tracker_addAllocatedBytes⚠

tracker_addInputBytes⚠

tracker_addOutputBytes⚠

tracker_addProcessingTimeNanos⚠

tracker_addSocket⚠

tracker_addVirtualProcessingDelay⚠

tracker_free⚠

tracker_heartbeat⚠

tracker_new⚠

tracker_removeAllocatedBytes⚠

tracker_removeSocket⚠

tracker_updateSocketInputBuffer⚠

tracker_updateSocketOutputBuffer⚠

tracker_updateSocketPeer⚠

Type Aliases

Address

An Address structure holds information used to identify nodes, allowing for easy extraction of both integer and string forms of an IP address as well as the string hostname associated with the IP. Address is an opaque structure and should only be accessed using the functions in this class.

CEmulatedTime

CSimulationTime

CompatSocket

CompatSocketObject

DNS

Epoll

Futex

GList

LegacyFile

LegacyFileCleanupFunc

LegacyFileCloseFunc

LegacyFileFreeFunc

LegacyFileFunctionTable

LegacySocket

NetworkInterface

Packet

PacketTCPHeader

ProtocolTCPFlags

ProtocolUDPFlags

RegularFile

ShadowSyscallNum

StatusArgumentFreeFunc

StatusCallbackFunc

StatusListener

StatusObjectFreeFunc

TCP

TCPCong

TCPCongDelete

TCPCongDuplicateAckEv

TCPCongFastRecovery

TCPCongHooks

TCPCongNameStr

TCPCongNewAckEv

TCPCongSSThresh

TCPCongTimeoutEv

Tracker

Trigger

TriggerObject

_CompatSocketTypes

_FileType

_LegacyFileType

_LogInfoFlags

_LogLevel

_PacketDeliveryStatusFlags

_ProtocolType

_StatusListenerFilter

_TriggerType

__blkcnt_t

__blksize_t

__dev_t

__fsblkcnt_t

__fsfilcnt_t

__fsword_t

__gid_t

__ino_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__s32

__s64

__suseconds_t

__syscall_slong_t

__time_t

__u16

__u32

__u64

__uid_t

_bindgen_ty_26

dev_t

gboolean

gchar

gid_t

gint

gpointer

gsize

gssize

guint

guint32

guint64

in_addr_t

in_port_t

linux___kernel_clock_t

linux___kernel_long_t

linux___kernel_pid_t

linux___kernel_timer_t

linux___kernel_uid32_t

linux___u32

linux_siginfo_t

linux_sigval_t

mode_t

nfds_t

off64_t

off_t

pid_t

sa_family_t

uid_t

Unions

_CompatSocketObject

_TriggerObject

linux___sifields

linux___sifields__bindgen_ty_5__bindgen_ty_1

linux_siginfo__bindgen_ty_1

linux_sigval