shadow_rs

Module cshadow

cbindgen:ignore

Re-exports

pub use self::_LegacyFileType as LegacyFileType;

pub use self::_ProtocolType as ProtocolType;

pub use self::_ProtocolTCPFlags as ProtocolTCPFlags;

pub use self::_PacketDeliveryStatusFlags as PacketDeliveryStatusFlags;

pub use self::_StatusListenerFilter as StatusListenerFilter;

pub use self::_FileType as FileType;

pub use self::_TriggerType as TriggerType;

Structs

CompatUntypedForeignPtr

SysCallCondition

TCPCongHooks_

TCPCong_

_Epoll

_Futex

_LegacyFileFunctionTable

_PacketSelectiveAckRange

_PacketSelectiveAcks

_PacketTCPHeader

_RegularFile

_StatusListener

_TCP

_Trigger

__BindgenBitfieldUnit

__BindgenOpaqueArray

If Bindgen could only determine the size and alignment of a type, it is represented like this.

__fsid_t

iovec

linux_dirent

linux_dirent64

pollfd

stat

statfs

statx

statx_timestamp

tcp_info

timespec

timeval

Constants

AFFINITY_UNINIT

CONFIG_HEADER_SIZE_TCP

CONFIG_MTU

CONFIG_PIPE_BUFFER_SIZE

GLIB_MAJOR_VERSION

GLIB_MICRO_VERSION

GLIB_MINOR_VERSION

SHADOW_SOMAXCONN

SYSCALL_IO_BUFSIZE

ShadowSyscallNum_SHADOW_SYSCALL_NUM_HOSTNAME_TO_ADDR_IPV4

ShadowSyscallNum_SHADOW_SYSCALL_NUM_INIT_MEMORY_MANAGER

ShadowSyscallNum_SHADOW_SYSCALL_NUM_SHADOW_YIELD

_FileType_FILE_TYPE_HOSTS

_FileType_FILE_TYPE_IN_MEMORY

_FileType_FILE_TYPE_LOCALTIME

_FileType_FILE_TYPE_NOTSET

_FileType_FILE_TYPE_RANDOM

_FileType_FILE_TYPE_REGULAR

_LegacyFileType_DT_EPOLL

_LegacyFileType_DT_FILE

_LegacyFileType_DT_NONE

_LegacyFileType_DT_TCPSOCKET

_PacketDeliveryStatusFlags_PDS_DESTROYED

_PacketDeliveryStatusFlags_PDS_INET_DROPPED

_PacketDeliveryStatusFlags_PDS_INET_SENT

_PacketDeliveryStatusFlags_PDS_NONE

_PacketDeliveryStatusFlags_PDS_RCV_INTERFACE_DROPPED

_PacketDeliveryStatusFlags_PDS_RCV_INTERFACE_RECEIVED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_BUFFERED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_DELIVERED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_DROPPED

_PacketDeliveryStatusFlags_PDS_RCV_SOCKET_PROCESSED

_PacketDeliveryStatusFlags_PDS_RCV_TCP_ENQUEUE_UNORDERED

_PacketDeliveryStatusFlags_PDS_RELAY_CACHED

_PacketDeliveryStatusFlags_PDS_RELAY_FORWARDED

_PacketDeliveryStatusFlags_PDS_ROUTER_DEQUEUED

_PacketDeliveryStatusFlags_PDS_ROUTER_DROPPED

_PacketDeliveryStatusFlags_PDS_ROUTER_ENQUEUED

_PacketDeliveryStatusFlags_PDS_SND_CREATED

_PacketDeliveryStatusFlags_PDS_SND_INTERFACE_SENT

_PacketDeliveryStatusFlags_PDS_SND_SOCKET_BUFFERED

_PacketDeliveryStatusFlags_PDS_SND_TCP_DEQUEUE_RETRANSMIT

_PacketDeliveryStatusFlags_PDS_SND_TCP_ENQUEUE_RETRANSMIT

_PacketDeliveryStatusFlags_PDS_SND_TCP_ENQUEUE_THROTTLED

_PacketDeliveryStatusFlags_PDS_SND_TCP_RETRANSMITTED

_ProtocolTCPFlags_PTCP_ACK

_ProtocolTCPFlags_PTCP_DUPACK

_ProtocolTCPFlags_PTCP_FIN

_ProtocolTCPFlags_PTCP_NONE

_ProtocolTCPFlags_PTCP_RST

_ProtocolTCPFlags_PTCP_SACK

_ProtocolTCPFlags_PTCP_SYN

_ProtocolType_PNONE

_ProtocolType_PTCP

_ProtocolType_PUDP

_StatusListenerFilter_SLF_ALWAYS

_StatusListenerFilter_SLF_NEVER

_StatusListenerFilter_SLF_OFF_TO_ON

_StatusListenerFilter_SLF_ON_TO_OFF

_TriggerType_TRIGGER_CHILD

_TriggerType_TRIGGER_DESCRIPTOR

_TriggerType_TRIGGER_FILE

_TriggerType_TRIGGER_FUTEX

_TriggerType_TRIGGER_NONE

Statics

SHADOW_FLAG_MASK⚠

TCP_CONG_RENO_NAME⚠

glib_major_version⚠

glib_micro_version⚠

glib_minor_version⚠

Functions

_syscallhandler_pollHelper⚠

_syscallhandler_readHelper⚠

_syscallhandler_validateLegacyFile⚠

_syscallhandler_writeHelper⚠

affinity_getGoodWorkerAffinity⚠

affinity_initPlatformInfo⚠

affinity_setProcessAffinity⚠

epoll_new⚠

futex_addListener⚠

futex_getAddress⚠

futex_getListenerCount⚠

futex_new⚠

futex_ref⚠

futex_removeListener⚠

futex_unref⚠

futex_unref_func⚠

futex_wake⚠

glib_check_version⚠

legacyfile_addFlags⚠

legacyfile_addListener⚠

legacyfile_adjustStatus⚠

legacyfile_clear⚠

legacyfile_close⚠

legacyfile_getEventSource⚠

legacyfile_getFlags⚠

legacyfile_getStatus⚠

legacyfile_getType⚠

legacyfile_init⚠

legacyfile_ref⚠

legacyfile_refWeak⚠

legacyfile_removeFlags⚠

legacyfile_removeListener⚠

legacyfile_setFlags⚠

legacyfile_supportsSaRestart⚠

legacyfile_unref⚠

legacyfile_unrefWeak⚠

legacysocket_addToInputBuffer⚠

legacysocket_addToOutputBuffer⚠

legacysocket_connectToPeer⚠

legacysocket_dropPacket⚠

legacysocket_getInputBufferLength⚠

legacysocket_getInputBufferSize⚠

legacysocket_getInputBufferSpace⚠

legacysocket_getOutputBufferLength⚠

legacysocket_getOutputBufferSize⚠

legacysocket_getOutputBufferSpace⚠

legacysocket_getPeerName⚠

legacysocket_getProtocol⚠

legacysocket_getSocketName⚠

legacysocket_getUnixPath⚠

legacysocket_isBound⚠

legacysocket_isFamilySupported⚠

legacysocket_isUnix⚠

legacysocket_peekNextInPacket⚠

legacysocket_peekNextOutPacket⚠

legacysocket_pullOutPacket⚠

legacysocket_pushInPacket⚠

legacysocket_receiveUserData⚠

legacysocket_removeFromInputBuffer⚠

legacysocket_removeFromOutputBuffer⚠

legacysocket_sendUserData⚠

legacysocket_setInputBufferSize⚠

legacysocket_setOutputBufferSize⚠

legacysocket_setPeerName⚠

legacysocket_setSocketName⚠

legacysocket_setUnix⚠

legacysocket_setUnixPath⚠

logger_get_global_start_time_micros⚠

process_parseArgStr⚠

process_parseArgStrFree⚠

process_vm_readv⚠

process_vm_writev⚠

regularfile_faccessat⚠

regularfile_fadvise⚠

regularfile_fallocate⚠

regularfile_fchmod⚠

regularfile_fchmodat⚠

regularfile_fchown⚠

regularfile_fchownat⚠

regularfile_fcntl⚠

regularfile_fgetxattr⚠

regularfile_flistxattr⚠

regularfile_flock⚠

regularfile_fremovexattr⚠

regularfile_fsetxattr⚠

regularfile_fstat⚠

regularfile_fstatat⚠

regularfile_fstatfs⚠

regularfile_fsync⚠

regularfile_ftruncate⚠

regularfile_futimesat⚠

regularfile_getFlagsAtOpen⚠

regularfile_getModeAtOpen⚠

regularfile_getOSBackedFD⚠

regularfile_getShadowFlags⚠

regularfile_getType⚠

regularfile_getdents⚠

regularfile_getdents64⚠

regularfile_ioctl⚠

regularfile_linkat⚠

regularfile_lseek⚠

regularfile_mkdirat⚠

regularfile_mknodat⚠

regularfile_new⚠

regularfile_open⚠

regularfile_openat⚠

regularfile_poll⚠

regularfile_pread⚠

regularfile_preadv⚠

regularfile_preadv2⚠

regularfile_pwrite⚠

regularfile_pwritev⚠

regularfile_pwritev2⚠

regularfile_read⚠

regularfile_readahead⚠

regularfile_readlinkat⚠

regularfile_renameat2⚠

regularfile_statx⚠

regularfile_symlinkat⚠

regularfile_sync_range⚠

regularfile_unlinkat⚠

regularfile_utimensat⚠

regularfile_write⚠

scanRpathForLib⚠

status_listener_compare⚠

statuslistener_new⚠

statuslistener_onStatusChanged⚠

statuslistener_ref⚠

statuslistener_setMonitorStatus⚠

statuslistener_unref⚠

syscallcondition_cancel⚠

syscallcondition_getActiveFile⚠

syscallcondition_getTimeout⚠

syscallcondition_new⚠

syscallcondition_newWithAbsTimeout⚠

syscallcondition_ref⚠

syscallcondition_setActiveFile⚠

syscallcondition_setTimeout⚠

syscallcondition_unref⚠

syscallcondition_waitNonblock⚠

syscallcondition_wakeupForSignal⚠

syscallhandler_creat⚠

syscallhandler_faccessat⚠

syscallhandler_faccessat2⚠

syscallhandler_fadvise64⚠

syscallhandler_fallocate⚠

syscallhandler_fchmod⚠

syscallhandler_fchmodat⚠

syscallhandler_fchmodat2⚠

syscallhandler_fchown⚠

syscallhandler_fchownat⚠

syscallhandler_fcntl⚠

syscallhandler_fdatasync⚠

syscallhandler_fgetxattr⚠

syscallhandler_flistxattr⚠

syscallhandler_flock⚠

syscallhandler_fremovexattr⚠

syscallhandler_fsetxattr⚠

syscallhandler_fstat⚠

syscallhandler_fstatfs⚠

syscallhandler_fsync⚠

syscallhandler_ftruncate⚠

syscallhandler_futimesat⚠

syscallhandler_getdents⚠

syscallhandler_getdents64⚠

syscallhandler_ioctl⚠

syscallhandler_linkat⚠

syscallhandler_lseek⚠

syscallhandler_mkdirat⚠

syscallhandler_mknodat⚠

syscallhandler_newfstatat⚠

syscallhandler_open⚠

syscallhandler_openat⚠

syscallhandler_poll⚠

syscallhandler_ppoll⚠

syscallhandler_pread64⚠

syscallhandler_preadv⚠

syscallhandler_preadv2⚠

syscallhandler_pselect6⚠

syscallhandler_pwrite64⚠

syscallhandler_pwritev⚠

syscallhandler_pwritev2⚠

syscallhandler_read⚠

syscallhandler_readahead⚠

syscallhandler_readlinkat⚠

syscallhandler_readv⚠

syscallhandler_renameat⚠

syscallhandler_renameat2⚠

syscallhandler_select⚠

syscallhandler_statx⚠

syscallhandler_symlinkat⚠

syscallhandler_sync_file_range⚠

syscallhandler_syncfs⚠

syscallhandler_unlinkat⚠

syscallhandler_utimensat⚠

syscallhandler_write⚠

syscallhandler_writev⚠

tcp_acceptServerPeer⚠

tcp_clearAllChildrenIfServer⚠

tcp_cong⚠

tcp_cong_reno_init⚠

tcp_disableReceiveBufferAutotuning⚠

tcp_disableSendBufferAutotuning⚠

tcp_enterServerMode⚠

tcp_getConnectionError⚠

tcp_getInfo⚠

tcp_getInputBufferLength⚠

tcp_getNotSentBytes⚠

tcp_getOutputBufferLength⚠

tcp_isListeningAllowed⚠

tcp_isValidListener⚠

tcp_networkInterfaceIsAboutToSendPacket⚠

tcp_new⚠

tcp_receiveUserData⚠

tcp_sendUserData⚠

tcp_setRustSocket⚠

tcp_shutdown⚠

tcp_updateServerBacklog⚠

tcpcong_nameStr⚠

Type Aliases

CEmulatedTime

CSimulationTime

Epoll

Futex

LegacyFile

LegacyFileCleanupFunc

LegacyFileCloseFunc

LegacyFileFreeFunc

LegacyFileFunctionTable

LegacySocket

PacketSelectiveAckRange

PacketSelectiveAcks

PacketTCPHeader

RegularFile

ShadowSyscallNum

StatusArgumentFreeFunc

StatusCallbackFunc

StatusListener

StatusObjectFreeFunc

TCP

TCPCong

TCPCongDelete

TCPCongDuplicateAckEv

TCPCongFastRecovery

TCPCongHooks

TCPCongNameStr

TCPCongNewAckEv

TCPCongSSThresh

TCPCongTimeoutEv

Trigger

TriggerObject

_FileType

_LegacyFileType

_PacketDeliveryStatusFlags

_ProtocolTCPFlags

_ProtocolType

_StatusListenerFilter

_TriggerType

__blkcnt_t

__blksize_t

__dev_t

__fsblkcnt_t

__fsfilcnt_t

__fsword_t

__gid_t

__ino_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__s32

__s64

__suseconds_t

__syscall_slong_t

__time_t

__u16

__u32

__u64

__uid_t

_bindgen_ty_26

dev_t

gboolean

gchar

gid_t

gint

gpointer

gsize

gssize

guint

guint32

in_addr_t

in_port_t

mode_t

nfds_t

off64_t

off_t

pid_t

sa_family_t

uid_t

Unions

_TriggerObject