shadow_rs

Module host

Process management and emulation of Linux behaviour.

This module contains the code responsible for starting the managed processes in a way that allows Shadow to intercept their syscalls. It also contains the emulation of Linux hosts, threads, processes, syscalls, files, network interfaces, etc.

Modules

context

This module provides several Context structs, intended to bundle together current relevant objects in the hierarchy.

cpu

descriptor

Linux file descriptors and file descriptions (equivalent to Linux struct files).

futex_table

host

An emulated Linux system.

managed_thread

A thread of a managed process.

memory_manager

Access and manage memory of a plugin process.

network

process

An emulated Linux process.

status_listener

Compatibility wrappers for C-style event listeners.

syscall

thread

An emulated Linux thread.

timer