Module shadow_rs::host::descriptor::listener

Types for emitting and subscribing to File events.

Structs

StateEventSource

A specified event source that passes a state and the changed bits to the function, but only if the monitored bits have changed and if the change the filter is satisfied.

Enums

StateListenerFilter

Type Aliases

StateListenHandle

Handles for event source listeners.