Module shadow_rs::host::memory_manager

Access and manage memory of a plugin process.

The starting point for the public API is MemoryManager. MemoryManager can be used to:

Directly read or write process memory

Obtain smart pointers (ProcessMemoryRef and ProcessMemoryRefMut) to process memory

Obtain cursors to process memory implementing std::io::Seek and either std::io::Read or std::io::Write (MemoryReaderCursor and MemoryWriterCursor)

For the MemoryManager to maintain a consistent view of the process’s address space, and for it to be able to enforce Rust’s safety requirements for references and sharing, all access to process memory must go through it. This includes servicing syscalls that modify the process address space (such as mmap).

Modules

export

memory_copier

memory_mapper

Structs

AllocdMem
Memory allocated by Shadow, in a remote address space.

MemoryManager
Provides accessors for reading and writing another process’s memory. When in use, any operation that touches that process’s memory must go through the MemoryManager to ensure soundness. See MemoryManager::new.

MemoryReaderCursor
An object implementing std::io::Read and std::io::Seek for a range of plugin memory.

MemoryWriterCursor
An object implementing std::io::Write and std::io::Seek for a range of plugin memory.

ProcessMemoryRef
An immutable reference to a slice of plugin memory. Implements Deref<[T]>, allowing, e.g.:

ProcessMemoryRefMut
A mutable reference to a slice of plugin memory. Implements DerefMut<[T]>, allowing, e.g.:

Enums

CopiedOrMapped

CopiedOrMappedMut

Functions

page_size

seek_helper
Shared implementation of seek for both MemoryReaderCursor and MemoryWriterCursor.