shadow_rs::host
Module memory_manager

Access and manage memory of a plugin process.

The starting point for the public API is MemoryManager. MemoryManager can be used to:

Directly read or write process memory

Obtain smart pointers (ProcessMemoryRef and ProcessMemoryRefMut) to process memory

Obtain cursors to process memory implementing std::io::Seek and either std::io::Read or std::io::Write (MemoryReaderCursor and MemoryWriterCursor)

For the MemoryManager to maintain a consistent view of the process’s address space, and for it to be able to enforce Rust’s safety requirements for references and sharing, all access to process memory must go through it. This includes servicing syscalls that modify the process address space (such as mmap).

Structs

AllocdMem

Memory allocated by Shadow, in a remote address space.

MemoryManager

Provides accessors for reading and writing another process’s memory.

MemoryReaderCursor

An object implementing std::io::Read and std::io::Seek for a range of plugin memory.

MemoryWriterCursor

An object implementing std::io::Write and std::io::Seek for a range of plugin memory.

ProcessMemoryRef

An immutable reference to a slice of plugin memory. Implements Deref<[T]>, allowing, e.g.:

ProcessMemoryRefMut

A mutable reference to a slice of plugin memory. Implements DerefMut<[T]>, allowing, e.g.: