Struct shadow_rs::host::memory_manager::MemoryManager

pub struct MemoryManager {
    memory_copier: MemoryCopier,
    memory_mapper: Option<MemoryMapper>,
    pid: Pid,
}

Provides accessors for reading and writing another process’s memory. When in use, any operation that touches that process’s memory must go through the MemoryManager to ensure soundness. See MemoryManager::new.

Fields
memory_copier: MemoryCopiermemory_mapper: Option<MemoryMapper>pid: Pid
Implementations

impl MemoryManager

pub unsafe fn new(pid: Pid) -> Self

SAFETY: pid’s memory must not be modified without holding an exclusive (mutable) reference to the returned MemoryManager. In Shadow we ensure this by:

Creating only one MemoryManager for a given process.

TODO: Not allowing any thread in the process to execute without holding a mutable reference to its MemoryManager.

Not directly modifying process memory via other techniques.

Assuming (!) nothing else concurrently modifies memory of the given process. (e.g. that some other process doesn’t start calling process_vm_writev to write to the process’s memory).

TODO: Validating that the process doesn’t have any shared memory mappings other than with Shadow or other simulated processes under Shadow’s control.

fn mapped_ref<'a, T: Pod + Debug>(
&'a self,
ptr: TypedPluginPtr<T>
) -> Option<&[T]>

fn mapped_mut<'a, T: Pod + Debug>(
&'a mut self,
ptr: TypedPluginPtr<T>
) -> Option<&mut [T]>

pub fn memory_ref<'a, T: Pod + Debug>(
&'a self,
ptr: TypedPluginPtr<T>
) -> Result<ProcessMemoryRef<'a, T>, Errno>

Returns a reference to the given memory, copying to a local buffer if the memory isn’t mapped into Shadow.

pub fn memory_ref_prefix<T: Pod + Debug>(
&self,
ptr: TypedPluginPtr<T>
) -> Result<ProcessMemoryRef<'_, T>, Errno>

Returns a reference to the memory from the beginning of the given pointer to the last address in the pointer that’s accessible. Useful for accessing string data of unknown size. The data is copied to a local buffer if the memory isn’t mapped into Shadow.

pub fn reader<'a>(&'a self, ptr: TypedPluginPtr<u8>) -> MemoryReaderCursor<'a>

Creates a std::io::Read accessor for the specified plugin memory. Useful for handing off the ability to read process memory to non-Shadow APIs, without copying it to local memory first.

pub fn read_vals<T: Pod + Debug, const N: usize>(
&self,
ptr: TypedPluginPtr<T>
) -> Result<[T; N], Errno>

Reads the memory into a local copy. memory_ref is potentially more efficient, but this is useful to avoid borrowing from the MemoryManager; e.g. when we still want to be able to access the data while also writing to process memory.

pub fn copy_from_ptr<T: Debug + Pod>(
&self,
dst: &mut [T],
src: TypedPluginPtr<T>
) -> Result<(), Errno>

Similar to read_vals, but saves a copy if you already have a dst to copy the data into.

fn copy_prefix_from_ptr<T: Debug + Pod>(
&self,
buf: &mut [T],
ptr: TypedPluginPtr<T>
) -> Result<usize, Errno>

pub fn copy_str_from_ptr<'a>(
&self,
dst: &'a mut [u8],
src: TypedPluginPtr<u8>
) -> Result<&'a CStr, Errno>

Copies a NULL-terminated string starting from the beginning of dst and contained completely within dst. Still works if some of dst isn’t readable, as long as a NULL-terminated-string is contained in the readable prefix.

If holding a reference to the MemoryManager for the lifetime of the string is acceptable, use memory_ref_prefix and ProcessMemoryRef::get_str to potentially avoid an extra copy.

pub fn memory_ref_mut<'a, T: Pod + Debug>(
&'a mut self,
ptr: TypedPluginPtr<T>
) -> Result<ProcessMemoryRefMut<'a, T>, Errno>

Returns a mutable reference to the given memory. If the memory isn’t mapped into Shadow, copies the data to a local buffer, which is written back into the process if and when the reference is flushed.

pub fn memory_ref_mut_uninit<'a, T: Pod + Debug>(
&'a mut self,
ptr: TypedPluginPtr<T>
) -> Result<ProcessMemoryRefMut<'a, T>, Errno>

Returns a mutable reference to the given memory. If the memory isn’t mapped into Shadow, just returns a buffer with unspecified contents, which will be written back into the process if and when the reference is flushed.

WARNING: If the reference is flushed without initializing its contents, the unspecified contents will be written back into process memory. This can be avoided by calling noflush on the reference.

pub fn copy_to_ptr<T: Pod + Debug>(
&mut self,
dst: TypedPluginPtr<T>,
src: &[T]
) -> Result<(), Errno>

Writes the memory from a local copy. If src doesn’t already exist, using memory_ref_mut_uninit and initializing the data in that reference saves a copy.

pub fn pid(&self) -> Pid

Which process’s address space this MemoryManager manages.

pub fn init_mapper(&mut self, thread: &mut impl Thread)

Initialize the MemoryMapper, allowing for more efficient access. Needs a running thread.

pub fn has_mapper(&self) -> bool

Whether the internal MemoryMapper has been initialized.

pub fn writer<'a>(
&'a mut self,
ptr: TypedPluginPtr<u8>
) -> MemoryWriterCursor<'a>

Create a write accessor for the specified plugin memory.

fn handle_brk(
&mut self,
thread: &mut impl Thread,
ptr: PluginPtr
) -> SyscallResult

fn do_mmap(
&mut self,
thread: &mut dyn Thread,
addr: PluginPtr,
length: usize,
prot: i32,
flags: i32,
fd: i32,
offset: i64
) -> SyscallResult

fn handle_munmap(
&mut self,
thread: &mut impl Thread,
addr: PluginPtr,
length: usize
) -> SyscallResult

fn do_munmap(
&mut self,
thread: &mut dyn Thread,
addr: PluginPtr,
length: usize
) -> Result<()>

fn handle_mremap(
&mut self,
thread: &mut impl Thread,
old_address: PluginPtr,
old_size: usize,
new_size: usize,
flags: i32,
new_address: PluginPtr
) -> SyscallResult

fn handle_mprotect(
&mut self,
thread: &mut impl Thread,
addr: PluginPtr,
size: usize,
prot: i32
) -> SyscallResult