Struct shadow_rs::host::memory_manager::MemoryManager

pub struct MemoryManager {
    shm_file: ShmFile,
    regions: IntervalMap<Region>,
    misses_by_path: HashMap<String, u32>,
    stack_copied: Interval,
    heap: Interval,
}

Manages the address-space for a plugin process.

The MemoryManager's primary purpose is to make plugin process's memory directly accessible to Shadow. It does this by tracking what regions of program memory in the plugin are mapped to what (analagous to /proc//maps), and remapping parts of the plugin's address space into a shared memory-file, which is also mapped into Shadow.

Shadow provides several methods for allowing Shadow to access the plugin's memory, such as get_readable_ptr. If the corresponding region of plugin memory is mapped into the shared memory file, the corresponding Shadow pointer is returned. If not, then, it'll fall back to (generally slower) Thread APIs.

For the MemoryManager to maintain consistent state, and to remap regions of memory it knows how to remap, Shadow must delegate handling of mman-related syscalls (such as mmap) to the MemoryManager via its handle_* methods.

Fields
shm_file: ShmFileregions: IntervalMap<Region>misses_by_path: HashMap<String, u32>stack_copied: Interval

The part of the stack that we've already remapped in the plugin. We initially mmap enough address space in Shadow to accomodate a large stack, but we only lazily allocate it. This is both to prevent wasting memory, and to handle that we can't map the current "working area" of the stack in thread-preload.

heap: Interval

The bounds of the heap. Note that before the plugin's first brk syscall this will be a zero-sized interval (though in the case of thread-preload that'll have already happened before we get control).

Implementations

impl MemoryManager

pub fn new(thread: &mut impl Thread) -> MemoryManager

fn unmap_mutations(&mut self, mutations: Vec<Mutation<Region>>)

Processes the mutations returned by an IntervalMap::insert or IntervalMap::clear operation. Each mutation describes a mapping that has been partly or completely overwritten (in the case of an insert) or cleared (in the case of clear).

Potentially:

Updates shadow_base on affected regions.

Deallocates space from shm_file.

Reclaims Shadow's address space via unmap.

When used on mutations after an insert, if the inserted region is to be mapped into shadow, be sure to call this before doing that mapping; otherwise we'll end up deallocating some or all of the space in that new mapping.

pub unsafe fn get_ref<T>(
&mut self,
thread: &mut impl Thread,
src: PluginPtr
) -> Result<&T>

Gets a reference to an object in the plugin's memory. SAFETY

The pointer must point to readable value of type T.

Returned ref mustn't be accessed after Thread runs again or flush is called.

pub unsafe fn get_mut_ref<T>(
&mut self,
thread: &mut impl Thread,
src: PluginPtr
) -> Result<&mut T>

Gets a mutable reference to an object in the plugin's memory. SAFETY

The pointer must point to a writeable value of type T.

Returned ref mustn't be accessed after Thread runs again or flush is called.

pub unsafe fn get_slice<T>(
&mut self,
thread: &mut impl Thread,
src: PluginPtr,
len: usize
) -> Result<&[T]>

Gets a slice of the plugin's memory. SAFETY

The pointer must point to a readable array of type T and at least size len.

Returned slice mustn't be accessed after Thread runs again or flush is called.

pub unsafe fn get_mut_slice<T>(
&mut self,
thread: &mut impl Thread,
src: PluginPtr,
len: usize
) -> Result<&mut [T]>

Gets a mutable slice of the plugin's memory. SAFETY

The pointer must point to a writeable array of type T and at least size len.

Returned slice mustn't be accessed after Thread runs again or flush is called.

pub fn handle_mmap(
&mut self,
thread: &mut impl Thread,
addr: PluginPtr,
length: usize,
prot: i32,
flags: i32,
fd: i32,
offset: i64
) -> Result<PluginPtr>

Shadow should delegate a plugin's call to mmap to this method. The caller is responsible for ensuring that fd is open and pointing to the right file in the plugin process.

Executes the actual mmap operation in the plugin, updates the MemoryManager's understanding of the plugin's address space, and in some cases remaps the given region into the MemoryManager's shared memory file for fast access. Currently only private anonymous mappings are remapped.

pub fn handle_munmap(
&mut self,
thread: &mut impl Thread,
addr: PluginPtr,
length: usize
) -> Result<()>

Shadow should delegate a plugin's call to munmap to this method.

Executes the actual mmap operation in the plugin, updates the MemoryManager's understanding of the plugin's address space, and unmaps the affected memory from Shadow if it was mapped in.

pub fn handle_mremap(
&mut self,
thread: &mut impl Thread,
old_address: PluginPtr,
old_size: usize,
new_size: usize,
flags: i32,
new_address: PluginPtr
) -> Result<PluginPtr>

Shadow should delegate a plugin's call to mremap to this method.

Executes the actual mremap operation in the plugin, updates the MemoryManager's understanding of the plugin's address space, and updates Shadow's mappings of that region if applicable.

pub fn handle_brk(
&mut self,
thread: &mut impl Thread,
ptr: PluginPtr
) -> Result<PluginPtr, i32>

Execute the requested brk and update our mappings accordingly. May invalidate outstanding pointers. (Rust won't allow mutable methods such as this one to be called with outstanding borrowed references).

pub fn handle_mprotect(
&mut self,
thread: &mut impl Thread,
addr: PluginPtr,
size: usize,
prot: i32
) -> Result<()>

Shadow should delegate a plugin's call to mprotect to this method.

Executes the actual mprotect operation in the plugin, updates the MemoryManager's understanding of the plugin's address space, and updates Shadow's mappings of that region if applicable.

Alternatively when Shadow maps a region it would always map everything to be both readable and writeable by Shadow, in which case we wouldn't need to worry about updating Shadow's protections when the plugin calls mprotect. However, mirroring the plugin's protection settings can help catch bugs earlier. Shadow should never have reason to access plugin memory in a way that the plugin itself can't.

fn extend_stack(&mut self, thread: &mut impl Thread, src: usize)

Extend the portion of the stack that we've mapped downward to include src.

This is carefuly designed not to invalidate any outstanding borrowed references or pointers, since otherwise a caller trying to marshall multiple syscall arguments might invalidate the first argument when marshalling the second. (While the Rust API currently prevents there from being any outstanding references, the C API does not).

fn get_mapped_ptr(
&mut self,
thread: &mut impl Thread,
src: PluginPtr,
n: usize
) -> Option<*mut c_void>

Get a raw pointer to the plugin's memory, if it's been remapped into Shadow (or if we can do so now, in the case of the stack).

fn inc_misses(&mut self, addr: PluginPtr)

Counts accesses where we had to fall back to the thread's (slow) apis.

unsafe fn get_readable_ptr(
&mut self,
thread: &mut impl Thread,
plugin_src: PluginPtr,
n: usize
) -> Result<*const c_void>

Get a readable pointer to the plugin's memory via mapping, or via the thread APIs. Never returns NULL.

unsafe fn get_writeable_ptr(
&mut self,
thread: &mut impl Thread,
plugin_src: PluginPtr,
n: usize
) -> Result<*mut c_void>

Get a writeable pointer to the plugin's memory via mapping, or via the thread APIs. Never returns NULL.

unsafe fn get_mutable_ptr(
&mut self,
thread: &mut impl Thread,
plugin_src: PluginPtr,
n: usize
) -> Result<*mut c_void>

Get a mutable pointer to the plugin's memory via mapping, or via the thread APIs. Never returns NULL.