Struct shadow_rs::host::process::Process

pub struct Process { /* private fields */ }

A simulated process.

Implementations

impl Process

pub fn borrow_as_runnable( &self ) -> Option<impl Deref<Target = RunnableProcess> + '_>

Borrows a reference to the internal RunnableProcess if self is runnable.

pub fn borrow_as_zombie( &self ) -> Option<impl Deref<Target = ZombieProcess> + '_>

Borrows a reference to the internal ZombieProcess if self is a zombie.

pub fn spawn( host: &Host, plugin_name: CString, plugin_path: &CStr, argv: Vec<CString>, envv: Vec<CString>, pause_for_debugging: bool, strace_logging_options: Option<FmtOptions>, expected_final_state: ProcessFinalState ) -> Result<RootedRc<RootedRefCell<Process>>>

Spawn a new process. The process will be runnable via Self::resume once it has been added to the Host’s process list.

pub fn id(&self) -> ProcessId

pub fn parent_id(&self) -> ProcessId

pub fn set_parent_id(&self, pid: ProcessId)

pub fn group_id(&self) -> ProcessId

pub fn set_group_id(&self, id: ProcessId)

pub fn session_id(&self) -> ProcessId

pub fn set_session_id(&self, id: ProcessId)

pub fn host_id(&self) -> HostId

pub fn dumpable(&self) -> SuidDump

Get process’s “dumpable” state, as manipulated by the prctl operations PR_SET_DUMPABLE and PR_GET_DUMPABLE.

pub fn set_dumpable(&self, val: SuidDump)

Set process’s “dumpable” state, as manipulated by the prctl operations PR_SET_DUMPABLE and PR_GET_DUMPABLE.

pub fn thread_group_leader_id(&self) -> ThreadId

pub fn resume(&self, host: &Host, tid: ThreadId)

Resume execution of tid (if it exists). Should only be called from Host::resume.

pub fn stop(&self, host: &Host)

Terminate the Process.

Should only be called from Host::free_all_applications.

pub fn signal( &self, host: &Host, current_thread: Option<&Thread>, siginfo_t: &siginfo_t )

See RunnableProcess::signal.

No-op if the self is a ZombieProcess.

pub fn name(&self) -> impl Deref<Target = str> + '_

pub fn plugin_name(&self) -> impl Deref<Target = str> + '_

pub fn memory_borrow_mut(&self) -> impl DerefMut<Target = MemoryManager> + '_

Deprecated wrapper for RunnableProcess::memory_borrow_mut

pub fn memory_borrow(&self) -> impl Deref<Target = MemoryManager> + '_

Deprecated wrapper for RunnableProcess::memory_borrow

pub fn strace_logging_options(&self) -> Option<FmtOptions>

Deprecated wrapper for RunnableProcess::strace_logging_options

pub fn with_strace_file<T>(&self, f: impl FnOnce(&mut File) -> T) -> Option<T>

Deprecated wrapper for RunnableProcess::with_strace_file

pub fn native_pid(&self) -> Pid

Deprecated wrapper for RunnableProcess::native_pid

pub fn realtime_timer_borrow(&self) -> impl Deref<Target = Timer> + '_

Deprecated wrapper for RunnableProcess::realtime_timer_borrow

pub fn realtime_timer_borrow_mut(&self) -> impl DerefMut<Target = Timer> + '_

Deprecated wrapper for RunnableProcess::realtime_timer_borrow_mut

pub fn first_live_thread_borrow( &self, root: &Root ) -> Option<impl Deref<Target = RootedRc<RootedRefCell<Thread>>> + '_>

Deprecated wrapper for RunnableProcess::first_live_thread_borrow

pub fn thread_borrow( &self, virtual_tid: ThreadId ) -> Option<impl Deref<Target = RootedRc<RootedRefCell<Thread>>> + '_>

Deprecated wrapper for RunnableProcess::thread_borrow

pub fn free_unsafe_borrows_flush(&self) -> Result<(), Errno>

Deprecated wrapper for RunnableProcess::free_unsafe_borrows_flush.

pub fn free_unsafe_borrows_noflush(&self)

Deprecated wrapper for RunnableProcess::free_unsafe_borrows_noflush.

pub fn physical_address( &self, vptr: ForeignPtr<()> ) -> ManagedPhysicalMemoryAddr

pub fn is_running(&self) -> bool

pub fn add_thread(&self, host: &Host, thread: RootedRc<RootedRefCell<Thread>>)

Deprecated wrapper for RunnableProcess::add_thread

pub fn shmem( &self ) -> impl Deref<Target = ShMemBlock<'static, ProcessShmem>> + '_

Deprecated wrapper for RunnableProcess::shmem

pub fn rusage(&self) -> rusage

Resource usage, as returned e.g. by the getrusage syscall.

pub fn exit_signal(&self) -> Option<Signal>

Signal that will be sent to parent process on exit. Typically Some(SIGCHLD).

pub fn current_working_dir(&self) -> impl Deref<Target = CString> + '_

pub fn update_for_exec( &mut self, host: &Host, tid: ThreadId, mthread: ManagedThread )

Update self to complete an exec syscall from thread tid, replacing the running managed process with mthread.