shadow_rs::host::syscall

Module formatter

Structs

NoArg

A marker type for indicating there are no types left in the syscall arguments.

SyscallArgsFmt

A formatting wrapper for six syscall arguments.

SyscallResultFmt

A formatting wrapper for the syscall result.

SyscallVal

A syscall argument or return value. It implements Display, and only reads memory and converts types when being formatted.

Enums

FmtOptions

StraceFmtMode

Traits

SyscallDisplay

Functions

log_syscall_simple

For logging unknown syscalls.

write_syscall

Format and write the syscall.