Trait shadow_rs::host::thread::Thread

pub trait Thread {
    fn native_syscall(
        &mut self, 
        n: i64, 
        args: &[SysCallReg]
    ) -> Result<SysCallReg>;

    fn get_readable_ptr(
        &mut self, 
        plugin_src: PluginPtr, 
        n: usize
    ) -> Result<*const c_void>;

    fn get_writeable_ptr(
        &mut self, 
        plugin_src: PluginPtr, 
        n: usize
    ) -> Result<*mut c_void>;

    fn get_mutable_ptr(
        &mut self, 
        plugin_src: PluginPtr, 
        n: usize
    ) -> Result<*mut c_void>;

    fn get_process_id(&self) -> u32;

    fn get_host_id(&self) -> u32;

    fn get_system_pid(&self) -> pid_t;

    fn flush(&mut self);

    fn native_munmap(
        &mut self, 
        ptr: PluginPtr, 
        size: usize
    ) -> Result<PluginPtr> { ... }

    fn native_mmap(
        &mut self, 
        addr: PluginPtr, 
        len: usize, 
        prot: i32, 
        flags: i32, 
        fd: i32, 
        offset: i64
    ) -> Result<PluginPtr> { ... }

    fn native_mremap(
        &mut self, 
        old_addr: PluginPtr, 
        old_len: usize, 
        new_len: usize, 
        flags: i32, 
        new_addr: PluginPtr
    ) -> Result<PluginPtr> { ... }

    fn native_mprotect(
        &mut self, 
        addr: PluginPtr, 
        len: usize, 
        prot: i32
    ) -> Result<()> { ... }

    fn native_open(
        &mut self, 
        pathname: PluginPtr, 
        flags: i32, 
        mode: i32
    ) -> Result<i32> { ... }

    fn native_close(&mut self, fd: i32) -> Result<()> { ... }

    fn native_brk(&mut self, addr: PluginPtr) -> Result<PluginPtr> { ... }

    fn malloc_plugin_ptr(&mut self, size: usize) -> Result<PluginPtr> { ... }

    fn free_plugin_ptr(&mut self, ptr: PluginPtr, size: usize) -> Result<()> { ... }
}

Required methods

fn native_syscall(&mut self, n: i64, args: &[SysCallReg]) -> Result<SysCallReg>

Have the plugin thread natively execute the given syscall.

fn get_readable_ptr(
&mut self,
plugin_src: PluginPtr,
n: usize
) -> Result<*const c_void>

Get a readable pointer to the plugin's memory. Never returns NULL.

fn get_writeable_ptr(
&mut self,
plugin_src: PluginPtr,
n: usize
) -> Result<*mut c_void>

Get a writeable pointer to the plugin's memory. Never returns NULL.

fn get_mutable_ptr(
&mut self,
plugin_src: PluginPtr,
n: usize
) -> Result<*mut c_void>

Get a mutable pointer to the plugin's memory. Never returns NULL. SAFETY

The specified memory region must be valid and writeable.

Returned pointer mustn't be accessed after Thread runs again or flush is called.

fn get_process_id(&self) -> u32

fn get_host_id(&self) -> u32

fn get_system_pid(&self) -> pid_t

fn flush(&mut self)

Flush (and invalidate) pointers previously returned by get_readable_ptr and get_writeable_ptr.

Loading content...
Provided methods

fn native_munmap(&mut self, ptr: PluginPtr, size: usize) -> Result<PluginPtr>

Natively execute munmap(2) on the given thread.

fn native_mmap(
&mut self,
addr: PluginPtr,
len: usize,
prot: i32,
flags: i32,
fd: i32,
offset: i64
) -> Result<PluginPtr>

Natively execute mmap(2) on the given thread.

fn native_mremap(
&mut self,
old_addr: PluginPtr,
old_len: usize,
new_len: usize,
flags: i32,
new_addr: PluginPtr
) -> Result<PluginPtr>

Natively execute mremap(2) on the given thread.

fn native_mprotect(
&mut self,
addr: PluginPtr,
len: usize,
prot: i32
) -> Result<()>

Natively execute mmap(2) on the given thread.

fn native_open(
&mut self,
pathname: PluginPtr,
flags: i32,
mode: i32
) -> Result<i32>

Natively execute open(2) on the given thread.

fn native_close(&mut self, fd: i32) -> Result<()>

Natively execute close(2) on the given thread.

fn native_brk(&mut self, addr: PluginPtr) -> Result<PluginPtr>

Natively execute brk(2) on the given thread.

fn malloc_plugin_ptr(&mut self, size: usize) -> Result<PluginPtr>

Allocates some space in the plugin's memory. Use get_writeable_ptr to write to it, and flush to ensure that the write is flushed to the plugin's memory.

fn free_plugin_ptr(&mut self, ptr: PluginPtr, size: usize) -> Result<()>

Frees a pointer previously returned by malloc_plugin_ptr

Loading content...