Module shadow_rs::utility::byte_queue

Structs

ByteQueue

A queue of bytes that supports reading and writing stream and/or packet data.

Enums

BytesWrapper

A wrapper type that holds either Bytes or BytesMut.

ChunkType

The types of data that are supported by the ByteQueue.