Struct shadow_rs::utility::byte_queue::ByteQueue

pub struct ByteQueue {
    bytes: LinkedList<ByteChunk>,
    unused_buffer: Option<BytesMut>,
    length: u64,
    default_chunk_capacity: usize,
}

A queue of bytes that supports reading and writing stream and/or packet data.

Both stream and packet data can be pushed onto the buffer and their order will be preserved. Data is stored internally as a linked list of chunks. Each chunk stores either stream or packet data. Consecutive stream data may be merged into a single chunk, but consecutive packets will always be contained in their own chunks.

To avoid memory copies when moving bytes from one ByteQueue to another, you can use pop_chunk() to remove a chunk from the queue, and use push_chunk() to add it to another queue.

Fields
bytes: LinkedList<ByteChunk>

The queued bytes.

unused_buffer: Option<BytesMut>

A pre-allocated buffer that can be used for new bytes.

length: u64

The number of bytes in the queue.

default_chunk_capacity: usize

The size of newly allocated chunks when storing stream data.

Implementations

impl ByteQueue

pub fn new(default_chunk_capacity: usize) -> Self

pub fn num_bytes(&self) -> u64

The number of bytes in the queue. If the queue has 0 bytes, it does not mean that the queue is empty since there may be 0-length packets in the queue.

pub fn has_bytes(&self) -> bool

Returns true if the queue has bytes.

pub fn has_chunks(&self) -> bool

Returns true if the queue has data/chunks, which may include packets with 0 bytes.

fn alloc_zeroed_buffer(&mut self, size: usize) -> BytesMut

pub fn push_stream<R: Read>(&mut self, src: R) -> Result<u64>

Push stream data onto the queue. The data may be merged into the previous stream chunk.

pub fn push_packet<R: Read>(&mut self, src: R, size: usize) -> Result<()>

Push packet data onto the queue in a single chunk. Exactly size bytes will be read into the packet.

pub fn push_chunk(
&mut self,
data: impl Into<BytesWrapper>,
chunk_type: ChunkType
) -> u64

Push a chunk of stream or packet data onto the queue.

pub fn pop<W: Write>(&mut self, dst: W) -> Result<(u64, Option<ChunkType>)>

Pop data from the queue. Only a single type of data will be popped per invocation. To read all data from the queue, you must call this method until the returned chunk type is None. Zero-length packets may be returned. If packet data is returned but dst did not have enough space, the remaining bytes in the packet will be dropped.

fn pop_stream<W: Write>(&mut self, dst: W) -> Result<u64>

fn pop_packet<W: Write>(&mut self, dst: W) -> Result<u64>

pub fn pop_chunk(&mut self, size_hint: usize) -> Option<(Bytes, ChunkType)>

Pop a single chunk of data from the queue. The size_hint argument is used to limit the number of bytes in the returned chunk iff the next chunk has stream data. If the returned chunk has packet data, the size_hint is ignored and the entire packet is returned.