Trait shadow_rs::utility::pod::Pod

pub unsafe trait Pod: Copy + 'static { }

Marker trait that the given type is Plain Old Data; i.e. that it is safe to interpret any pattern of bits as a value of this type.

This is notably not true for many Rust types. e.g. interpreting the integer value 2 as a rust bool is undefined behavior.

We require Copy to also rule out anything that implements Drop.

References are inherently non-Pod, so we can require a ’static lifetime.